% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggPathwayView.R
\docType{methods}
\name{arrangePathview}
\alias{arrangePathview}
\title{Kegg pathway view and arrange grobs on page}
\usage{
arrangePathview(
  genelist,
  pathways = c(),
  top = 4,
  ncol = 2,
  title = NULL,
  sub = NULL,
  organism = "hsa",
  output = ".",
  path.archive = ".",
  kegg.native = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{genelist}{a data frame with columns of ENTREZID, Control and Treatment. The columns
of Control and Treatment represent gene score in Control and Treatment sample.}

\item{pathways}{character vector, the KEGG pathway ID(s), usually 5 digit, may also
include the 3 letter KEGG species code.}

\item{top}{integer, specifying how many top enriched pathways to be visualized.}

\item{ncol}{integer, specifying how many column of figures to be arranged in each page.}

\item{title}{optional string, or grob.}

\item{sub}{optional string, or grob.}

\item{organism}{character, either the kegg code, scientific name or the common name of
the target species. This applies to both pathway and gene.data or cpd.data. When KEGG
ortholog pathway is considered, species="ko". Default species="hsa", it is equivalent
to use either "Homo sapiens" (scientific name) or "human" (common name).}

\item{output}{Path to save plot to.}

\item{path.archive}{character, the directory of KEGG pathway data file (.xml) and image file
(.png). Users may supply their own data files in the same format and naming convention
 of KEGG's (species code + pathway id, e.g. hsa04110.xml, hsa04110.png etc) in this
 directory. Default kegg.dir="." (current working directory).}

\item{kegg.native}{logical, whether to render pathway graph as native KEGG graph (.png)
or using graphviz layout engine (.pdf). Default kegg.native=TRUE.}

\item{verbose}{Boolean}
}
\value{
plot on the current device
}
\description{
Kegg pathway view and arrange grobs on page.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFluteRadioScreen"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
colnames(dd)[2:3] = c("Control", "Treatment")
# arrangePathview(dd, c("hsa00534"), title=NULL, sub=NULL, organism="hsa")

}
\author{
Wubing Zhang
}
