\name{BRAF_mle.gene_summary}
\alias{BRAF_mle.gene_summary}
\docType{data}
\title{
Gene summary object of BRAF dataset
}
\description{
The gene summary results generated by running MAGeCK MLE on Feng Zhang's BRAF CRISPR screening dataset.
}
\usage{data("BRAF_mle.gene_summary")}
\format{
  A data frame with 17419 observations on the 26 variables.
}

\examples{
data(BRAF_mle.gene_summary)
head(BRAF_mle.gene_summary)
}
\keyword{datasets}
