\name{BRAF_rra.gene_summary}
\alias{BRAF_rra.gene_summary}
\docType{data}
\title{
Gene summary object of BRAF dataset
}
\description{
The gene summary results generated by running MAGeCK RRA on Feng Zhang's BRAF CRISPR screening dataset.
}
\usage{data("BRAF_rra.gene_summary")}
\format{
  A data frame with 17140 observations on 14 variables.
}

\examples{
head(BRAF_rra.gene_summary)
}
\keyword{datasets}
