% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsistencyView.R
\docType{methods}
\name{ConsistencyView}
\alias{ConsistencyView}
\title{Visualize the estimate cell cycle compared to control.}
\usage{
ConsistencyView(
  beta,
  ctrlname,
  treatname,
  main = NULL,
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{beta}{Data frame, which has columns of \code{ctrlname} and other samples.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the name of treatment samples.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Estimate cell cycle time in different samples by linear fitting of beta scores.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
ConsistencyView(dd, ctrlname = "dmso", treatname = "plx")

}
\author{
Wubing Zhang
}
