% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrView.R
\docType{methods}
\name{CorrView}
\alias{CorrView}
\title{Visualize the correlation between two object}
\usage{
CorrView(gg, x, y, smoothMethod = "lm", main = NULL, xlab = NULL,
  ylab = NULL, filename = NULL, width = 5, height = 4, ...)
}
\arguments{
\item{gg}{A data frame.}

\item{x}{A character, indicating column (in countSummary) of x-axis.}

\item{y}{A character, indicating column (in countSummary) of y-axis.}

\item{smoothMethod}{A character, indicating fill color of all bars.}

\item{main}{A charater, specifying the figure title.}

\item{xlab}{A character, specifying the title of x-axis.}

\item{ylab, }{A character, specifying the title of y-axis.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Visualize the correlation between two object
}
\examples{
gg = data.frame(x = rnorm(50), y = rnorm(50))
CorrView(gg, x="x", y="y")

}
