\name{countsummary}
\alias{countsummary}
\docType{data}
\title{
  Count summary data generated by running MAGeCK count
}
\description{
  The summary of QC values at count level
}
\usage{
  data("countsummary")
}
\format{
  A data frame with 4 observations on 13 variables.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25476604}
}

\examples{
  data("countsummary")
  head(countsummary)
}
\keyword{datasets}

