% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityView.R
\docType{methods}
\name{DensityDiffView}
\alias{DensityDiffView}
\title{Density plot}
\usage{
DensityDiffView(
  dat,
  ctrlname = "Control",
  treatname = "Treatment",
  main = NULL,
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{dat}{A data frame.}

\item{ctrlname}{A character, specifying the control samples.}

\item{treatname}{A character, specifying the treatment samples.}

\item{main}{A character, specifying title.}

\item{filename}{A character, specifying a file name to create on disk.
Set filename to be "NULL", if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the distribution of score differences between treatment and control.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
# Density plot of beta score deviation between control and treatment
DensityDiffView(dd, ctrlname = "Pmel1_Ctrl", treatname = "Pmel1")


}
\author{
Wubing Zhang
}
