% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityView.R
\docType{methods}
\name{DensityView}
\alias{DensityView}
\title{Density plot for gene beta scores in Control and Treatment}
\usage{
DensityView(beta, ctrlname = "Control", treatname = "Treatment",
  main = NULL, xlab = "Beta Score", filename = NULL)
}
\arguments{
\item{beta}{data frame, which has columns of 'Gene', \code{ctrlname} and \code{treatname}.}

\item{ctrlname}{a character, specifying the name of control sample.}

\item{treatname}{a character, specifying the name of treatment sample.}

\item{main}{as in 'plot'.}

\item{xlab}{as in 'plot'.}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the density of gene beta scores in two samples.
}
\note{
See the vignette for an example of density plot for beta score deviation.
Note that the source code of \code{DensityView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::DensityView}
or \code{getMethod("DensityView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/DensityView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
#Density plot of all gene beta scores in control and treatment
DensityView(dd)


}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{KeggPathwayView}}

\code{\link{ViolinView}}   \code{\link{SquareView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{MAView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
