% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepmapConfidenceIntervals.R
\docType{methods}
\name{DepmapConfidenceIntervals}
\alias{DepmapConfidenceIntervals}
\title{Load processed Depmap data}
\usage{
DepmapConfidenceIntervals(dd)
}
\value{
data.frame with confidence intervals added
}
\description{
Load processed Depmap data
}
\author{
Robin Jugas
}
