% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich.DAVID}
\alias{enrich.DAVID}
\title{enrichment analysis by DAVID}
\usage{
enrich.DAVID(gene, universe = NULL, david.user = "ma.tongji@gmail.com",
  minGSSize = 2, maxGSSize = 500, annotation = "GOTERM_BP_FAT",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", qvalueCutoff = 0.2)
}
\arguments{
\item{pAdjustMethod}{one of "BH" and "bonferroni"}
}
\description{
enrichment analysis by DAVID
}
