% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GOstats.R
\docType{methods}
\name{enrich.GOstats}
\alias{enrich.GOstats}
\alias{enrichGOstats}
\title{Do enrichment analysis using GOstats}
\usage{
enrich.GOstats(gene, universe = NULL, type = c("KEGG", "BP", "MF", "CC"),
  organism = "hsa", pvalueCutoff = 0.25, pAdjustMethod = "BH")
}
\arguments{
\item{gene}{A character vector, specifying the genelist to do enrichment analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{type}{Geneset category for testing, KEGG(default).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using GOstats method
}
\examples{
data(geneList, package = "DOSE")
genes <- names(geneList)[1:100]
enrichRes <- enrich.GOstats(genes, type="BP")
head(enrichRes@result)

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
