% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{GSEA}
\usage{
enrich.GSE(geneList, keytype = "Entrez", type = "MsigDB",
  organism = "hsa", pvalueCutoff = 0.05, pAdjustMethod = "BH",
  limit = c(3, 50), gmtpath = NA)
}
\arguments{
\item{geneList}{A order ranked numeric vector with geneid as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, "MsigDB" (default), "KEGG" or "Others" (`gmtpath` is required).}

\item{organism}{A character, specifying organism, only 'human' is available.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A integer vector of length two, specifying the limit of geneset size.}

\item{gmtpath}{The path to gmt file.}
}
\value{
A enrichResult instance.
}
\description{
A universal gene set enrichment analysis tools
}
\examples{
data(geneList, package = "DOSE")
\dontrun{
    enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
    head(enrichRes@result)
}

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
