% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.ORT.R
\docType{methods}
\name{enrich.ORT}
\alias{enrich.ORT}
\alias{enrichORT}
\title{Do enrichment analysis using over-representation test}
\usage{
enrich.ORT(geneList, universe = NULL, keytype = "Entrez", type = "MsigDB",
  organism = "hsa", pvalueCutoff = 0.05, pAdjustMethod = "BH",
  limit = c(3, 50), gmtpath = NA)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, "MsigDB" (default), "KEGG", "BP", "CC", "MF", "DO",
"MKEGG", "NCG" or "Others" (`gmtpath` is required).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A integer vector of length two, specifying the limit of geneset size.}

\item{gmtpath}{The path to gmt file.}
}
\value{
A enrichedResult instance.
}
\description{
Do enrichment analysis using over-representation test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
enrichedRes <- enrich.ORT(genes)
head(enrichedRes@result)

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.GSE}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
