% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichAnalyzer.R
\docType{methods}
\name{EnrichAnalyzer}
\alias{EnrichAnalyzer}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
EnrichAnalyzer(geneList, keytype = "Entrez",
  type = "CORUM+GOBP+GOMF+GOCC+KEGG", method = "ORT",
  organism = "hsa", pvalueCutoff = 0.25, limit = c(3, 80),
  universe = NULL, filter = TRUE, gmtpath = NA)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'CORUM', 'CPX' (ComplexPortal),
'GOBP', 'GOMF', 'GOCC', 'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID',
or any combination of them (e.g. 'GOBP+GOMF+CORUM'), or 'All' (all categories).}

\item{method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), and "HGT"(HyperGemetric test).}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{filter}{Boolean, specifying whether filter out redundancies from the enrichment results.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
\code{enrichRes} is an enrichResult instance.
}
\description{
Enrichment analysis
}
\examples{
data(geneList, package = "DOSE")
keggA = EnrichAnalyzer(geneList, method = "GSEA")
head(keggA@result)

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrich.HGT}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Feizhen Wu
}
