% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedGeneView.R
\docType{methods}
\name{EnrichedGeneView}
\alias{EnrichedGeneView}
\title{Visualize enriched pathways and genes in those pathways}
\usage{
EnrichedGeneView(enrichment, geneList, rank_by = "p.adjust", top = 5,
  bottom = 0, keytype = "Symbol", gene_cutoff = c(-log2(1.5),
  log2(1.5)), custom_gene = NULL, charLength = 40, filename = NULL,
  width = 7, height = 5, ...)
}
\arguments{
\item{enrichment}{A data frame of enrichment result or an \code{enrichResult} object.}

\item{geneList}{A numeric geneList used in enrichment anlaysis.}

\item{rank_by}{"p.adjust", "pvalue", or "NES", specifying the indices for ranking pathways.}

\item{top}{An integer, specifying the number of positively enriched terms to show.}

\item{bottom}{An integer, specifying the number of negatively enriched terms to show.}

\item{keytype}{"Entrez" or "Symbol".}

\item{gene_cutoff}{A two-length numeric vector, specifying cutoff for genes to show.}

\item{custom_gene}{A character vector (gene names), customizing genes to show.}

\item{charLength}{Integer, specifying max length of enriched term name to show as coordinate lab.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Visualize enriched pathways and genes in those pathways
}
\examples{
data(geneList, package = "DOSE")
enrichRes <- enrich.GSE(geneList)
EnrichedGeneView(enrichment=slot(enrichRes, "result"), geneList, keytype = "Entrez")
}
\author{
Wubing Zhang
}
