% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\docType{methods}
\name{enrichment_analysis}
\alias{enrichment_analysis}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
enrichment_analysis(geneList, method = "HGT", keytype = "Entrez",
  type = "KEGG", organism = "hsa", pvalueCutoff = 0.25,
  limit = c(3, 50), universe = NULL, main = NULL)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), and "HGT"(HyperGemetric test).}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'CORUM', 'CPX' (ComplexPortal),
'GOBP', 'GOMF', 'GOCC', 'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID',
or any combination of them (e.g. 'GOBP+GOMF+CORUM'), or 'All' (all categories).}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{main}{Same as 'title' in 'plot'.}
}
\value{
A list, including two items, \code{gridPlot} and \code{enrichRes}. \code{gridPlot} is
a ggplot object, and \code{enrichRes} is a enrichResult instance.
}
\description{
Enrichment analysis
}
\examples{
data(geneList, package = "DOSE")
keggA = enrichment_analysis(geneList, method = "GSEA", type = "KEGG")
print(keggA$gridPlot)

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrich.HGT}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Feizhen Wu
}
