% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBADsgRNA.R
\docType{methods}
\name{filterBADsgRNA}
\alias{filterBADsgRNA}
\title{Filter bad sgRNA guides}
\usage{
filterBADsgRNA(fileCountNormalized, Squareview_table)
}
\arguments{
\item{fileCountNormalized}{}

\item{Squareview_table}{}
}
\value{
DF
}
\description{
Go over sgRNA counts and filter & mark the unsatisfying ones
}
\author{
Robin Jugas, Matej Jasik
}
