\name{gene2path_hsa}
\alias{gene2path_hsa}
\docType{data}
\title{
mapping of genes and pathways
}
\description{
mapping of genes to pathways for hsa
}
\usage{data("gene2path_hsa")}
\format{
  A data frame with 28245 observations on the following 2 variables.
  \describe{
    \item{\code{EntrezID}}{a character vector}
    \item{\code{PathwayID}}{a character vector}
  }
}

\keyword{datasets}
