\name{gene2path_mmu}
\alias{gene2path_mmu}
\docType{data}
\title{
mapping of genes and pathways
}
\description{
mapping of genes to pathways for mmu
}
\usage{gene2path_mmu}
\format{
  A data frame with 2 variables.
  \describe{
    \item{\code{EntrezID}}{a character vector}
    \item{\code{PathwayID}}{a character vector}
  }
}

\keyword{datasets}
