\name{GeneID_Convert_hsa}
\alias{GeneID_Convert_hsa}
\docType{data}
\title{
gene name mapping data for hsa
}
\description{
mapping of ENTREZID and SYMBOL for hsa
}
\usage{GeneID_Convert_hsa}
\format{
  A data frame with 115954 observations on 2 variables.
    \describe{
      \item{\code{ENTREZID}}{a character vector}
      \item{\code{SYMBOL}}{a character vector}
  }
}

\keyword{datasets}
