\name{GeneID_Convert_mmu}
\alias{GeneID_Convert_mmu}
\docType{data}
\title{
gene name mapping data for mmu
}
\description{
mapping of ENTREZID and SYMBOL for mmu
}
\usage{GeneID_Convert_mmu}
\format{
  A data frame with 205941 observations on the following 2 variables.
  \describe{
    \item{\code{ENTREZID}}{a character vector}
    \item{\code{SYMBOL}}{a character vector}
  }
}

\keyword{datasets}
