% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustView.R
\docType{methods}
\name{getCols}
\alias{getCols}
\title{Map values to colors}
\usage{
getCols(x, palette = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{palette}{diverge, rainbow, sequential}
}
\value{
A vector of colors corresponding to input vector.
}
\description{
Map values to colors
}
\author{
Wubing Zhang
}
