% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{getOrg}
\alias{getOrg}
\title{Determine the gene annotation package.}
\usage{
getOrg(organism, update = FALSE)
}
\arguments{
\item{organism}{Character, KEGG species code, or the common species name, used to determine
the gene annotation package. For all potential values check: data(bods); bods. Default org="hsa",
and can also be "human" (case insensitive).}

\item{update}{Boolean, indicating whether download recent annotation from NCBI.}
}
\value{
A list containing three elements:
\item{organism}{species}
\code{pkg}{annotation package name}
\code{Symbol_Entrez}{a data frame, mapping between gene symbol and entrez id}
}
\description{
Determine the gene annotation package. for specific organism
}
\examples{
ann = getOrg("human")
print(ann$pkg)

}
\author{
Wubing Zhang
}
