% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{getOrtAnn}
\alias{getOrtAnn}
\title{Retreive reference orthologs annotation.}
\usage{
getOrtAnn(fromOrg = "mmu", toOrg = "hsa", update = FALSE)
}
\arguments{
\item{fromOrg}{Character, hsa (default), bta, cfa, mmu, ptr, rno, ssc are optional.}

\item{toOrg}{Character, hsa (default), bta, cfa, mmu, ptr, rno, ssc are optional.}

\item{update}{Boolean, indicating whether download recent annotation from NCBI.}
}
\value{
A data frame.
}
\description{
Retreive reference orthologs annotation.
}
\examples{
ann = getOrtAnn("hsa", "mmu")
head(ann)

}
\author{
Wubing Zhang
}
