% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{gsGetter}
\alias{gsGetter}
\title{Extract pathway annotation from GMT file.}
\usage{
gsGetter(gmtpath = NULL, type = "All", limit = c(0, Inf),
  organism = "hsa")
}
\arguments{
\item{gmtpath}{The path to customized gmt file.}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (PID, KEGG, REACTOME, BIOCARTA, C2CP), GO (GOBP, GOCC, GOMF),
Complex (CORUM, CPX), c1, c2, c3, c4, c6, c7, HALLMARK. It also allows any
combination of them (e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets to load.}

\item{organism}{'hsa' or 'mmu'.}
}
\value{
A three-column data frame.
}
\description{
Extract pathway annotation from GMT file.
}
\examples{
gene2path = gsGetter(type = "REACTOME+CORUM")
head(gene2path)

}
\author{
Wubing Zhang
}
