% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapView.R
\docType{methods}
\name{HeatmapView}
\alias{HeatmapView}
\alias{heatmapview}
\title{Calculate the similarity between samples and plot heatmap}
\usage{
HeatmapView(beta, method = "pearson",
  color = rev(colorRampPalette(RColorBrewer::brewer.pal(7, "RdYlBu"))(200)),
  breaks = NA, cluster_rows = TRUE, cluster_cols = TRUE, legend = TRUE,
  main = NA, fontsize = 10, display_numbers = TRUE, filename = NA,
  width = NA, height = NA, ...)
}
\arguments{
\item{beta}{data frame or matrix, in which each column represents one sample.}

\item{method}{character, One of "pearson", "kendall", "spearman", "euclidean", "maximum",
"manhattan", "canberra", "binary", or "minkowski".}

\item{color}{the same as that in pheatmap}

\item{breaks}{the same as that in pheatmap}

\item{cluster_rows}{the same as that in pheatmap}

\item{cluster_cols}{the same as that in pheatmap}

\item{legend}{the same as that in pheatmap}

\item{main}{the same as that in pheatmap}

\item{fontsize}{the same as that in pheatmap}

\item{display_numbers}{the same as that in pheatmap}

\item{filename}{the same as that in pheatmap}

\item{width}{the same as that in pheatmap}

\item{height}{the same as that in pheatmap}

\item{...}{other parameters in pheatmap}
}
\value{
The same as pheatmap
}
\description{
Calculate the similarity between samples and plot heatmap
}
\note{
See the vignette for an example of HeatmapView
Note that the source code of \code{HeatmapView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::HeatmapView}
or \code{getMethod("HeatmapView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/HeatmapView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
dd = ReadBeta(MLE_Data, organism="hsa")
dd = dd[,3:ncol(dd)]
HeatmapView(dd, method = "pearson")

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\author{
Wubing Zhang
}
