% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncorporateDepmap.R
\docType{methods}
\name{loadDepmap}
\alias{loadDepmap}
\alias{LoadDepmap}
\title{Load processed Depmap data}
\usage{
LoadDepmap(dependency = -0.5)
}
\value{
A list including two elements, one is the Depmap CRISPR data, and the
other is the sample annotation data.
}
\description{
Load processed Depmap data
}
\examples{
\dontrun{
  depmapDat = LoadDepmap()
}
}
\author{
Wubing Zhang
}
