% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginfo.R
\docType{methods}
\name{loginfo}
\alias{loginfo}
\alias{log}
\title{Log message}
\usage{
loginfo(msg)
}
\arguments{
\item{msg}{a character, specifying log message.}
}
\value{
print log message on stdout.
}
\description{
Write log message on console
}
\note{
The source can be found by typing \code{MAGeCKFlute:::loginfo}
or \code{getMethod("loginfo")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/loginfo.R}
Users should find it easy to customize this function.
}
\examples{
loginfo("Test the loginfo function ...")

}
\author{
Binbin Wang
}
