% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRatesView.R
\docType{methods}
\name{MapRatesView}
\alias{MapRatesView}
\title{View mapping ratio}
\usage{
MapRatesView(
  countSummary,
  Label = "Label",
  Reads = "Reads",
  Mapped = "Mapped",
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{countSummary}{A data frame, which contains columns of
`Label`, `Reads`, and `Mapped`}

\item{Label}{A character, indicating column (in countSummary) of sample names.}

\item{Reads}{A character, indicating column (in countSummary) of total reads.}

\item{Mapped}{A character, indicating column (in countSummary) of mapped reads.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
View mapping ratio of each sample
}
\examples{
file4 = file.path(system.file("extdata", package = "MAGeCKFluteRadioScreen"),
"testdata/countsummary.txt")
countsummary = read.delim(file4, check.names = FALSE)
MapRatesView(countsummary)

}
\author{
Wubing Zhang
}
