\name{mle.gene_summary}
\alias{mle.gene_summary}
\docType{data}
\title{
  Gene summary table in MAGeCK MLE results
}
\description{
  The gene summary results generated by running MAGeCK MLE on CRISPR screens.
}
\usage{
  data("mle.gene_summary")
}
\format{
  A data frame.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/26673418}
}
\examples{
  data("mle.gene_summary")
  head(mle.gene_summary)
}
\keyword{datasets}
