\name{MLE_Data}
\alias{MLE_Data}
\docType{data}
\title{
  Gene summary table in MAGeCK MLE results
}
\description{
  The gene summary results generated by running MAGeCK MLE on CRISPR screens.
}
\usage{
  data("MLE_Data")
}
\format{
  A data frame with 17141 observations on the 26 variables.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/26673418}
}
\examples{
  data("MLE_Data")
  head(MLE_Data)
}
\keyword{datasets}
