% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedView.R
\docType{methods}
\name{noEnrichPlot}
\alias{noEnrichPlot}
\title{Blank figure}
\usage{
noEnrichPlot(main = "No enriched terms")
}
\arguments{
\item{main}{The title of figure.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Blank figure
}
\author{
Wubing Zhang
}
