% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBeta.R
\docType{methods}
\name{NormalizeBeta}
\alias{NormalizeBeta}
\alias{normalizebeta}
\title{Normalize gene beta scores}
\usage{
NormalizeBeta(
  beta,
  id = 1,
  method = "cell_cycle",
  posControl = NULL,
  samples = NULL,
  org = "hsa"
)
}
\arguments{
\item{beta}{Data frame.}

\item{id}{An integer specifying the column of gene.}

\item{method}{Character, one of 'cell_cycle'(default) and 'loess'.
or character string giving the name of the table column containing the gene names.}

\item{posControl}{A character vector, specifying a list of positive control genes.}

\item{samples}{Character vector, specifying the sample names in \emph{beta} columns.
If NULL (default), take all \emph{beta} columns as samples.}

\item{org}{"hsa", "mmu", "bta", "cfa", "ptr", "rno", or "ssc" indicating the organism.}
}
\value{
A data frame with same format as input data \emph{beta}.
}
\description{
Two normalization methods are available. \code{cell_cycle} method normalizes gene beta scores
 based on positive control genes in CRISPR screening. \code{loess} method normalizes gene
 beta scores using loess.
}
\details{
In CRISPR screens, cells treated with different conditions (e.g., with or without drug)
may have different proliferation rates. So it's necessary to normalize the proliferation rate
based on defined positive control genes among samples. After normalization, the beta scores are
comparable across samples. \code{loess} is another optional normalization method, which is used
to normalize array data before.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFluteRadioScreen"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
\dontrun{
  #Cell Cycle normalization
  dd_essential = NormalizeBeta(dd, method="cell_cycle", org = "mmu")
  head(dd_essential)
}
#Optional loess normalization (not recommended)
dd_loess = NormalizeBeta(dd, method="loess")
head(dd_loess)

}
\author{
Wubing Zhang
}
