% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncorporateDepmap.R
\docType{methods}
\name{OmitCommonEssential}
\alias{OmitCommonEssential}
\title{Omit common essential genes based on depmap data}
\usage{
OmitCommonEssential(
  dd,
  symbol = "id",
  lineages = "All",
  cell_lines = NULL,
  dependency = -0.5
)
}
\arguments{
\item{dd}{A data frame.}

\item{symbol}{A character, specifying the column name of gene symbols in the data frame.}

\item{lineages}{A character vector, specifying the lineages for selecting essential genes.}

\item{cell_lines}{A character vector, specifying cell lines for selecting essential genes.}

\item{dependency}{A numeric, specifying the threshold for selecting essential genes.}
}
\value{
A data frame.
}
\description{
Omit common essential genes based on depmap data
}
\examples{
\dontrun{
  file1 = file.path(system.file("extdata", package = "MAGeCKFluteRadioScreen"),
                    "testdata/rra.gene_summary.txt")
  gdata = ReadRRA(file1)
  dim(gdata)
  rra.omit = OmitCommonEssential(gdata)
  dim(rra.omit)
}

}
\author{
Wubing Zhang
}
