\name{pathways_hsa}
\alias{pathways_hsa}
\docType{data}
\title{
mapping of pathwayID to pathway name
}
\description{
mapping of pathwayID to pathway name for hsa
}
\usage{pathways_hsa}
\format{
  A data frame with 321 observations on the following 2 variables.
  \describe{
    \item{\code{PathwayID}}{a character vector}
    \item{\code{PathwayName}}{a character vector}
  }
}
\keyword{datasets}
