\name{pathways_mmu}
\alias{pathways_mmu}
\docType{data}
\title{
mapping of pathwayID to pathway name
}
\description{
mapping of pathwayID to pathway name for mmu
}
\usage{pathways_mmu}
\format{
  A data frame with 317 observations on the following 2 variables.
  \describe{
    \item{\code{PathwayID}}{a character vector}
    \item{\code{PathwayName}}{a character vector}
  }
}
\keyword{datasets}
