% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankView.R
\docType{methods}
\name{RankView}
\alias{RankView}
\alias{rankview}
\title{Rank plot}
\usage{
RankView(
  rankdata,
  genelist = NULL,
  decreasing = TRUE,
  top = 5,
  bottom = 5,
  cutoff = 2,
  main = NULL,
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{rankdata}{A numeric vector, with gene as names.}

\item{genelist}{A character vector, specifying genes to be labeled.}

\item{decreasing}{Boolean, specifying the order of genes to plot.}

\item{top}{Integer, specifying number of positive genes to be labeled.}

\item{bottom}{Integer, specifying number of negative genes to be labeled.}

\item{cutoff}{One numeric value indicating the fold of standard deviation used as cutoff;
two number vector, such as c(-1, 1), specifying the exact cutoff for selecting top genes.}

\item{main}{A character, specifying title.}

\item{filename}{A character, specifying a file name to create on disk.
Set filename to be "NULL", if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other available parameters in the function 'geom_text_repel'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Draw the score and rank of genes on a scatter plot.
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFluteRadioScreen"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
rankdata = gdata$Score
names(rankdata) = gdata$id
RankView(rankdata)

}
\author{
Wubing Zhang
}
