% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBeta.R
\docType{methods}
\name{ReadBeta}
\alias{ReadBeta}
\alias{readbeta}
\title{Read gene beta scores}
\usage{
ReadBeta(gene_summary)
}
\arguments{
\item{gene_summary}{A data frame or a file path to gene summary file generated by MAGeCK-MLE.}
}
\value{
A data frame, whose first column is Gene and other columns are comparisons.
}
\description{
Read gene beta scores from file or data frame
}
\examples{
data(mle.gene_summary)
dd = ReadBeta(mle.gene_summary)
head(dd)

}
\author{
Wubing Zhang
}
