% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCountAgregate.R
\docType{methods}
\name{ReadCountAgregate}
\alias{ReadCountAgregate}
\title{Aggregate read counts and collapse into gene clusters}
\usage{
ReadCountAgregate(normalised_count, Squareview_table)
}
\arguments{
\item{normalised_count}{table of normalized counts}

\item{Squareview_table}{Squareview table}
}
\value{
DF
}
\description{
Go over sgRNA counts and filter & mark the unsatisfying ones
}
\author{
Matej Jasik
}
