% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{goplot}
\alias{gseaplot}
\alias{gseaplot2}
\alias{heatplot}
\alias{ridgeplot}
\alias{GSEA}
\alias{enricher}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{clusterProfiler}{\code{\link[clusterProfiler]{enricher}}, \code{\link[clusterProfiler]{GSEA}}}

  \item{enrichplot}{\code{\link[enrichplot:reexports]{cnetplot}}, \code{\link[enrichplot]{dotplot}}, \code{\link[enrichplot]{emapplot}}, \code{\link[enrichplot]{goplot}}, \code{\link[enrichplot]{gseaplot}}, \code{\link[enrichplot]{gseaplot2}}, \code{\link[enrichplot]{heatplot}}, \code{\link[enrichplot]{ridgeplot}}}
}}

