% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{retrieve_gs}
\alias{retrieve_gs}
\title{Update genesets from source database}
\usage{
retrieve_gs(type = c("KEGG", "REACTOME", "CORUM"), organism = "hsa")
}
\arguments{
\item{type}{A vector of databases, such as KEGG, REACTOME, CORUM.}

\item{organism}{'hsa' or 'mmu'.}
}
\value{
save data to local library.
}
\description{
Update genesets from source database
}
\author{
Wubing Zhang
}
