\name{rra.gene_summary}
\alias{rra.gene_summary}
\docType{data}
\title{
  Gene summary data generated by running MAGeCK RRA
}
\description{
  The gene summary results generated by running MAGeCK on CRISPR screens.
}
\usage{
  data("rra.gene_summary")
}
\format{
  A data frame.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25476604}
}

\examples{
  data("rra.gene_summary")
  head(rra.gene_summary)
}
\keyword{datasets}
