\name{rra.sgrna_summary}
\alias{rra.sgrna_summary}
\docType{data}
\title{
  sgRNA summary data generated by running MAGeCK RRA
}
\description{
  The sgRNA summary results generated by running `mageck test` on CRISPR screens.
}
\usage{data("rra.sgrna_summary")}
\format{
  A data frame.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25476604}
}
\examples{
data(rra.sgrna_summary)
head(rra.sgrna_summary)
}
\keyword{datasets}
