\name{RRA_Data}
\alias{RRA_Data}
\docType{data}
\title{
  Gene summary data generated by running MAGeCK RRA
}
\description{
  The gene summary results generated by running MAGeCK on CRISPR screens.
}
\usage{
  data("RRA_Data")
}
\format{
  A data frame with 17140 observations on 14 variables.
}
\references{
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25494202}
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25476604}
}

\examples{
  data("RRA_Data")
  head(RRA_Data)
}
\keyword{datasets}
