% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterView.R
\docType{methods}
\name{ScatterView}
\alias{ScatterView}
\title{Scatter plot}
\usage{
ScatterView(
  data,
  x = "x",
  y = "y",
  label = 0,
  model = c("none", "ninesquare", "volcano", "rank")[1],
  x_cut = NULL,
  y_cut = NULL,
  slope = 1,
  intercept = NULL,
  auto_cut = FALSE,
  auto_cut_x = auto_cut,
  auto_cut_y = auto_cut,
  auto_cut_diag = auto_cut,
  groups = NULL,
  group_col = NULL,
  groupnames = NULL,
  label.top = TRUE,
  top = 0,
  toplabels = NULL,
  display_cut = FALSE,
  color = NULL,
  shape = 16,
  size = 1,
  alpha = 0.6,
  main = NULL,
  xlab = x,
  ylab = y,
  xlim = NULL,
  ylim = NULL,
  legend.position = "None",
  ...
)
}
\arguments{
\item{data}{Data frame.}

\item{x}{A character, specifying the x-axis.}

\item{y}{A character, specifying the y-axis.}

\item{label}{An integer or a character specifying the column used as the label, default value is 0 (row names).}

\item{model}{One of "none" (default), "ninesquare", "volcano", and "rank".}

\item{x_cut}{An one or two-length numeric vector, specifying the cutoff used for x-axis.}

\item{y_cut}{An one or two-length numeric vector, specifying the cutoff used for y-axis.}

\item{slope}{A numberic value indicating slope of the diagonal cutoff.}

\item{intercept}{A numberic value indicating intercept of the diagonal cutoff.}

\item{auto_cut}{Boolean or numeric, specifying how many standard deviation will be used as cutoff.}

\item{auto_cut_x}{Boolean or numeric, specifying how many standard deviation will be used as cutoff on x-axis.}

\item{auto_cut_y}{Boolean or numeric, specifying how many standard deviation will be used as cutoff on y-axis}

\item{auto_cut_diag}{Boolean or numeric, specifying how many standard deviation will be used as cutoff on diagonal.}

\item{groups}{A character vector specifying groups. Optional groups include "top", "mid", "bottom",
"left", "center", "right", "topleft", "topcenter", "topright", "midleft", "midcenter",
"midright", "bottomleft", "bottomcenter", "bottomright".}

\item{group_col}{A vector of colors for specified groups.}

\item{groupnames}{A vector of group names to show on the legend.}

\item{label.top}{Boolean, specifying whether label top hits.}

\item{top}{Integer, specifying the number of top terms in the groups to be labeled.}

\item{toplabels}{Character vector, specifying terms to be labeled.}

\item{display_cut}{Boolean, indicating whether display the dashed line of cutoffs.}

\item{color}{A character, specifying the column name of color in the data frame.}

\item{shape}{A character, specifying the column name of shape in the data frame.}

\item{size}{A character, specifying the column name of size in the data frame.}

\item{alpha}{A numeric, specifying the transparency of the dots.}

\item{main}{Title of the figure.}

\item{xlab}{Title of x-axis}

\item{ylab}{Title of y-axis.}

\item{xlim}{x-axis range}

\item{ylim}{y-axis range}

\item{legend.position}{Position of legend, "none", "right", "top", "bottom", or
a two-length vector indicating the position.}

\item{...}{Other available parameters in function 'geom_text_repel'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Scatter plot supporting groups.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
ScatterView(dd, x = "Pmel1_Ctrl", y = "Pmel1", label = "Gene",
auto_cut = 1, groups = "topright", top = 5, display_cut = TRUE)
ScatterView(dd, x = "Pmel1_Ctrl", y = "Pmel1", label = "Gene",
auto_cut = 2, model = "ninesquare", top = 5, display_cut = TRUE)

}
\author{
Wubing Zhang
}
