% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selector.R
\docType{methods}
\name{Selector}
\alias{Selector}
\title{Select signatures from candidate list (according to the consistence in most samples).}
\usage{
Selector(mat, cutoff = 0, type = "<", select = 0.8)
}
\arguments{
\item{mat}{A matrix, each row is candidates (genes), each column is samples.}

\item{cutoff}{Numeric, specifying the cutoff to define the signatures.}

\item{type}{Character, ">" or "<".}

\item{select}{Numeric, specifying the proportion of samples in which signature is selected.}
}
\value{
An list containing two elements, the first is the selected signature and the second is a ggplot object.
}
\description{
Select signatures from candidate list (according to the consistence in most samples).
}
\examples{
mat = matrix(rnorm(1000*30), 1000, 30)
rownames(mat) = paste0("Gene", 1:1000)
colnames(mat) = paste0("Sample", 1:30)
hits = Selector(mat, select = 0.68)
print(hits$p)

}
