% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFluteMLE.R
\docType{methods}
\name{subFluteMLE}
\alias{subFluteMLE}
\title{Downstream analysis based on MAGeCK-MLE result}
\arguments{
\item{dd}{A data frame.}

\item{ctrlname}{A character vector, specifying the names of control samples in \code{dd}.}

\item{treatname}{A character vector, specifying the names of treatment samples in \code{dd}.}

\item{organism}{"hsa" or "mmu".}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{top}{An integer, specifying number of top selected genes to be labeled in rank figure.}

\item{bottom}{An integer, specifying number of bottom selected genes to be labeled in rank figure.}

\item{interestGenes}{A character vector, specifying interested genes to be labeled in rank figure.}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{pvalueCutoff}{A numeric, specifying pvalue cutoff of enrichment analysis, default 1.}

\item{adjust}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{enrich_kegg}{One of "ORT" and "HGT"(HyperGemetric test).}

\item{posControl}{A character vector, specifying a list of positive control gene symbols.}
}
\description{
Identify essential genes from MAGeCK MLE results
}
\author{
Wubing Zhang
}
