% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoView.R
\docType{methods}
\name{VolcanoView}
\alias{VolcanoView}
\title{Volcano View}
\usage{
VolcanoView(
  df,
  x = "logFC",
  y = "adj.P.Val",
  Label = NA,
  top = 5,
  topnames = NULL,
  x_cutoff = log2(1.5),
  y_cutoff = 0.05,
  mycolour = c("gray80", "#e41a1c", "#377eb8"),
  alpha = 0.6,
  force = 0.1,
  main = NULL,
  xlab = "log2FC",
  ylab = "-log10(FDR)",
  filename = NULL,
  width = 4,
  height = 2.5,
  ...
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A character, specifying the x-axis in Volcanno figure, 'logFC' (default).}

\item{y}{A character, specifying the y-axis in Volcanno figure, 'adj.P.Val' (default).
log10 transformation will be done automatically.}

\item{Label}{A character, specifying dots to be labeled on the figure.}

\item{top}{An integer, specifying the number of top significant genes to be labeled.}

\item{topnames}{A character vector, indicating positive/negative controls to be labeled.}

\item{x_cutoff}{Numeric, specifying cutoff of the x-axis.}

\item{y_cutoff}{Numeric, specifying cutoff of the y-axis.}

\item{mycolour}{A color vector, specifying colors of non-significant, significantly up and down-regulated genes.}

\item{alpha}{Numeric, parameter in ggplot.}

\item{force}{Numeric, Parameter for geom_text_repel. Force of repulsion between overlapping text labels.}

\item{main}{A character, specifying title.}

\item{xlab}{A character, specifying title of x-axis.}

\item{ylab}{A character, specifying title of y-axis.}

\item{filename}{A character, specifying a file name to create on disk. Set filename to be "NULL",
if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Volcano plot for differential analysis.
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
VolcanoView(gdata, x = "Score", y = "FDR", Label = "id")

}
\author{
Wubing Zhang
}
