% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\docType{methods}
\name{writeGMT}
\alias{writeGMT}
\title{Write GMT file}
\usage{
writeGMT(gene2path, gmtfile)
}
\arguments{
\item{gene2path}{A data frame. The columns should be Gene, Pathway ID, and Pathway Name.}

\item{gmtfile}{Path to gmt file.}
}
\value{
Output gmt file to local folder.
}
\description{
write data frame to a gmt file
}
\examples{
gene2path = gsGetter(type = "Complex")
# writeGMT(gene2path, "Protein_complex.gmt")

}
\author{
Wubing Zhang
}
