% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggPathwayView.R
\docType{methods}
\name{arrangePathview}
\alias{arrangePathview}
\title{Kegg pathway view and arrange grobs on page}
\usage{
arrangePathview(genelist, pathways = c(), organism = "hsa",
  view_allpath = FALSE, title = "Group A",
  sub = "Negative control normalized", output = ".", path.archive = ".",
  kegg.native = TRUE)
}
\arguments{
\item{genelist}{a data frame with columns of ENTREZID, Control and Treatment. The columns
of Control and Treatment represent gene score in Control and Treatment sample.}

\item{pathways}{character vector, the KEGG pathway ID(s), usually 5 digit, may also
include the 3 letter KEGG species code.}

\item{organism}{character, either the kegg code, scientific name or the common name of
the target species. This applies to both pathway and gene.data or cpd.data. When KEGG
ortholog pathway is considered, species="ko". Default species="hsa", it is equivalent
to use either "Homo sapiens" (scientific name) or "human" (common name).}

\item{view_allpath}{boolean, specifying whether view all pathways. Default view_allpath='FALSE',
and only plot top 4 enriched pathways.}

\item{title}{optional string, or grob.}

\item{sub}{optional string, or grob.}

\item{output}{Path to save plot to.}

\item{path.archive}{character, the directory of KEGG pathway data file (.xml) and image file
(.png). Users may supply their own data files in the same format and naming convention
 of KEGG's (species code + pathway id, e.g. hsa04110.xml, hsa04110.png etc) in this
 directory. Default kegg.dir="." (current working directory).}

\item{kegg.native}{logical, whether to render pathway graph as native KEGG graph (.png)
or using graphviz layout engine (.pdf). Default kegg.native=TRUE.}
}
\description{
Kegg pathway view and arrange grobs on page.
}
\examples{
\dontrun{
  data(MLE_Data)
  # Read beta score from gene summary table in MAGeCK MLE results
  dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
  arrangePathview(dd, pathways_hsa$PathwayID[1:2], title=NULL, sub=NULL, organism="hsa")
}

}
\seealso{
\code{\link{KeggPathwayView}}
}
\author{
Wubing Zhang
}
