% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchRemove.R
\docType{methods}
\name{BatchRemove}
\alias{BatchRemove}
\title{Batch effect removal}
\usage{
BatchRemove(mat, batchMat, log2trans = FALSE, positive = FALSE)
}
\arguments{
\item{mat}{Matrix, or a file path of data.}

\item{batchMat}{Matrix like data object or a file path of batch table,
in which the first two columns are `Samples`(matched colname of mat) and `Batch`, and
remaining columns should be Covariates.}

\item{log2trans}{Boolean, specifying whether do log2 transition before batch removal.}

\item{positive}{Boolean, specifying whether all values should be positive.}
}
\value{
A list contrains two objects, including \code{data} and \code{p}.
}
\description{
Batch effect removal
}
\examples{
edata = matrix(c(rnorm(2000, 5), rnorm(2000, 8)), 1000)
colnames(edata) = paste0("s", 1:4)
batchMat = data.frame(sample = colnames(edata), batch = rep(1:2, each = 2))
edata1 = BatchRemove(edata, batchMat)
print(edata1$p)

}
\seealso{
\code{\link[sva]{ComBat}}
}
\author{
Wubing Zhang
}
