% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchRemove.R
\docType{methods}
\name{BatchRemove}
\alias{BatchRemove}
\title{Batch effect removal}
\usage{
BatchRemove(mat, batchMat, log2trans = FALSE)
}
\arguments{
\item{mat}{A data frame, each row is a gene, and each column is a sample.}

\item{batchMat}{A data frame, the first column should be `Samples`(matched colnames of mat)
and the second column is `Batch`. The remaining columns could be Covariates.}

\item{log2trans}{Boolean, specifying whether do logarithmic transformation before batch removal.}
}
\value{
A list contrains two objects, including \code{data} and \code{p}.
}
\description{
Batch effect removal
}
\examples{
edata = matrix(c(rnorm(2000, 5), rnorm(2000, 8)), 1000)
colnames(edata) = paste0("s", 1:4)
batchMat = data.frame(sample = colnames(edata), batch = rep(1:2, each = 2))
edata1 = BatchRemove(edata, batchMat)
print(edata1$p)

}
\seealso{
\code{\link[sva]{ComBat}}
}
\author{
Wubing Zhang
}
