% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchRemove.R
\docType{methods}
\name{BatchRemove}
\alias{BatchRemove}
\alias{batchremove}
\title{Batch effect removal}
\usage{
BatchRemove(mat, batchMat, cluster = TRUE, log2trans = FALSE, ...)
}
\arguments{
\item{mat}{data frame or matrix, in which each column represents one sample.}

\item{batchMat}{data frame or matrix, which has three columns of Samples matched colname of mat, Batch, and Covariate.}

\item{cluster}{boolean, specifing whether do cluster analysis before and after batch removal}

\item{log2trans}{boolean, specifing whether do log2 transition before batch removal}

\item{...}{other parameters in pheatmap}
}
\value{
matrix of data after batch removal.
}
\description{
Remove batch effect
}
\note{
See the vignette for an example of BatchRemove
Note that the source code of \code{BatchRemove} is very simple.
The source can be found by typing \code{MAGeCKFlute:::BatchRemove}
or \code{getMethod("BatchRemove")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/BatchRemove.R}
Users should find it easy to customize this function.
}
\examples{
beta = ReadBeta(MLE_Data, organism="hsa")
batchMat = data.frame(samples = c("D7_R1", "D7_R2", "PLX7_R1", "PLX7_R2"),
                      batch = c(1,2,1,2), cov = c(1,1,2,2))
res = BatchRemove(beta, batchMat, cluster = FALSE)
}
\seealso{
\code{\link[sva]{ComBat}}
}
\author{
Wubing Zhang
}
