% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchRemove.R
\docType{methods}
\name{BatchRemove}
\alias{BatchRemove}
\title{Batch effect removal}
\usage{
BatchRemove(mat, batchMat, log2trans = FALSE, positive = FALSE)
}
\arguments{
\item{mat}{Matrix, or a file path of data.}

\item{batchMat}{Matrix like data object or a file path of batch table,
in which the first two columns are `Samples`(matched colname of mat) and `Batch`, and
remaining columns should be Covariates.}

\item{log2trans}{Boolean, specifying whether do log2 transition before batch removal.}

\item{positive}{Boolean, specifying whether all values should be positive.}
}
\value{
A list contrains two objects, including \code{data} and \code{p}.
}
\description{
Batch effect removal
}
\examples{
data(bladderdata, package = "bladderbatch")
dat <- bladderEset[1:50,]
pheno = pData(dat)
edata = exprs(dat)
batchMat = pheno[, c("sample", "batch", "cancer")]
batchMat$sample = rownames(batchMat)
edata1 = BatchRemove(edata, batchMat)

}
\seealso{
\code{\link[sva]{ComBat}}
}
\author{
Wubing Zhang
}
