% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchRemove.R
\docType{methods}
\name{BatchRemove}
\alias{BatchRemove}
\alias{batchremove}
\title{Batch effect removal}
\usage{
BatchRemove(mat, batchMat, log2trans = FALSE)
}
\arguments{
\item{mat}{Matrix like data object, or a file path of data.}

\item{batchMat}{Matrix like data object or a file path of batch table, which has at least two columns,
including Samples(matched colname of mat) and Batch. It can have the third column, which should be Covariate.}

\item{log2trans}{Boolean, specifying whether do log2 transition before batch removal.}
}
\value{
A list contrains two objects, including \code{data} and \code{p}.
}
\description{
Remove batch effect
}
\note{
See the vignette for an example of BatchRemove
Note that the source code of \code{BatchRemove} is very simple.
The source can be found by typing \code{MAGeCKFlute:::BatchRemove}
or \code{getMethod("BatchRemove")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/BatchRemove.R}
Users should find it easy to customize this function.
}
\examples{
beta = ReadBeta(MLE_Data, organism="hsa")
batchMat = data.frame(samples = c("D7_R1", "D7_R2", "PLX7_R1", "PLX7_R2"),
                      batch = c("bat1","bat2","bat1","bat2"), cov = c(1,1,2,2))
res = BatchRemove(beta, batchMat)

}
\seealso{
\code{\link[sva]{ComBat}}
}
\author{
Wubing Zhang
}
