% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCycleView.R
\docType{methods}
\name{CellCycleView}
\alias{CellCycleView}
\alias{CellCycle,MAGeCKFlute-method}
\title{Plot of linear fitting lines for beta scores in Control and Treatment sample}
\usage{
CellCycleView(beta, ctrlname = "Control", main = NULL, filename = NULL)
}
\arguments{
\item{beta}{data frame, which has columns of 'Gene', \code{ctrlname} and other samples.}

\item{ctrlname}{character vector, specifying the name of control sample.}

\item{main}{as in 'plot'.}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
no output.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Estimate cell cycle time in different samples by linear fitting of beta scores, and plot fitting lines,
in which x-axis is control beta score and y-axis is beta score of all samples.
}
\note{
See the vignette for an example of cell cycle estimate and view.
Note that the source code of \code{CellCycleFit} is very simple.
The source can be found by typing \code{MAGeCKFlute:::CellCycleView}
or \code{getMethod("CellCycleView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/CellCycleView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")[,-2]
CellCycleView(dd, ctrlname = c("D7_R1", "D7_R2"))

}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{DensityView}}

\code{\link{ViolinView}}   \code{\link{SquareView}}

\code{\link{KeggPathwayView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{MAView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
