% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsistencyView.R
\docType{methods}
\name{ConsistencyView}
\alias{ConsistencyView}
\title{Visualize the estimate cell cycle compared to control.}
\usage{
ConsistencyView(
  dat,
  ctrlname,
  treatname,
  main = NULL,
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{dat}{A data frame.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the names of treatment samples.}

\item{main}{A character, specifying title.}

\item{filename}{A character, specifying a file name to create on disk.
Set filename to be "NULL", if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Estimate cell cycle time in different samples by linear fitting of beta scores.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
ConsistencyView(dd, ctrlname = "Pmel1_Ctrl", treatname = "Pmel1")

}
\author{
Wubing Zhang
}
