% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutoffCalling.R
\docType{methods}
\name{CutoffCalling}
\alias{CutoffCalling}
\title{Call cutoff}
\usage{
CutoffCalling(d, scale = FALSE)
}
\arguments{
\item{d}{A numeric vector.}

\item{scale}{Boolean or numeric, whether scale cutoff to whole genome level, or how many standard deviation will be used as cutoff.}
}
\value{
A numeric value.
}
\description{
Calculate standard deviation as cutoff for a numeric vector
}
