% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutoffCalling.R
\docType{methods}
\name{CutoffCalling}
\alias{CutoffCalling}
\title{Call cutoff}
\usage{
CutoffCalling(d, scale = FALSE)
}
\arguments{
\item{d}{a numeric vector.}

\item{scale}{boolean or numeric, whether scale cutoff to whole genome level, or how many standard deviation will be used as cutoff.}
}
\value{
a numeric value
}
\description{
Calculate standard deviation as cutoff for a numeric vector
}
\note{
See the vignette for an example of cutoff calling.
Note that the source code of \code{CutoffCalling} is very simple.
The source can be found by typing \code{MAGeCKFlute:::CutoffCalling}
or \code{getMethod("CutoffCalling")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/CutoffCalling.R}
Users should find it easy to customize this function.
}
\examples{
cutoff=CutoffCalling(rnorm(18000), scale = TRUE)
print(cutoff)

}
\author{
Wubing Zhang
}
