% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterView.R
\docType{methods}
\name{CutoffCalling}
\alias{CutoffCalling}
\title{Quantile of normal distribution.}
\usage{
CutoffCalling(d, scale = 1)
}
\arguments{
\item{d}{A numeric vector.}

\item{scale}{Boolean or numeric, specifying how many standard deviation will be used as cutoff.}
}
\value{
A numeric value.
}
\description{
Compute cutoff from a normal-distributed vector.
}
\examples{
CutoffCalling(rnorm(10000))
}
