% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityDiffView.R
\docType{methods}
\name{DensityDiffView}
\alias{DensityDiffView}
\title{Density plot for beta score deviation between Control and Treatment}
\usage{
DensityDiffView(beta, ctrlname = "Control", treatname = "Treatment",
  main = NULL, filename = NULL)
}
\arguments{
\item{beta}{data frame, which has columns of 'Gene', \code{ctrlname} and \code{treatname}.}

\item{ctrlname}{a character, specifying the name of control sample.}

\item{treatname}{a character, specifying the name of treatment sample.}

\item{main}{as in 'plot'.}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
no output.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the density of beta score deviation between two samples.
}
\note{
See the vignette for an example of density plot for beta score deviation.
Note that the source code of \code{DensityDiffView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::DensityDiffView}
or \code{getMethod("DensityDiffView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/DensityDiffView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
# Density plot of beta score deviation between control and treatment
DensityDiffView(dd)


}
\seealso{
\code{\link{KeggPathwayView}}   \code{\link{DensityView}}

\code{\link{ViolinView}}   \code{\link{SquareView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{MAView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
