% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityDiffView.R
\docType{methods}
\name{DensityDiffView}
\alias{DensityDiffView}
\title{Density plot}
\usage{
DensityDiffView(beta, ctrlname = "Control", treatname = "Treatment",
  main = NULL, filename = NULL, width = 5, height = 4, ...)
}
\arguments{
\item{beta}{Data frame, including \code{ctrlname} and \code{treatname} as columns.}

\item{ctrlname}{A character, specifying the name of control sample.}

\item{treatname}{A character, specifying the name of treatment sample.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the density of beta score deviations.
}
\examples{
data(mle.gene_summary)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(mle.gene_summary)
# Density plot of beta score deviation between control and treatment
DensityDiffView(dd, ctrlname = "dmso", treatname = "plx")


}
\author{
Wubing Zhang
}
