% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityView.R
\docType{methods}
\name{DensityView}
\alias{DensityView}
\title{Density plot for gene beta scores in Control and Treatment}
\usage{
DensityView(
  beta,
  samples = NULL,
  main = NULL,
  xlab = "Beta Score",
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{beta}{Data frame, including \code{samples} as columns.}

\item{samples}{Character, specifying sample names in \code{beta}.}

\item{main}{As in 'plot'.}

\item{xlab}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the density of gene beta scores in two samples.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
DensityView(dd, samples=c("dmso", "plx"))
#or
DensityView(dd[, c("dmso", "plx")])

}
\seealso{
\code{\link{ViolinView}}
}
\author{
Wubing Zhang
}
