% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityView.R
\docType{methods}
\name{DensityView}
\alias{DensityView}
\title{Density plot}
\usage{
DensityView(
  dat,
  samples = NULL,
  main = NULL,
  xlab = "Score",
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{dat}{A data frame.}

\item{samples}{A character vector, specifying columns in \code{dat} for plotting.}

\item{main}{A character, specifying title.}

\item{xlab}{A character, specifying title of x-axis.}

\item{filename}{A character, specifying a file name to create on disk.
Set filename to be "NULL", if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the distribution of numeric vectors with the same length.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
DensityView(dd, samples=c("dmso", "plx"))
#or
DensityView(dd[, c("dmso", "plx")])

}
\seealso{
\code{\link{ViolinView}}
}
\author{
Wubing Zhang
}
