% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityView.R
\docType{methods}
\name{DensityView}
\alias{DensityView}
\title{Density plot for gene beta scores in Control and Treatment}
\usage{
DensityView(beta, samples = NULL, main = NULL, xlab = "Beta Score",
  filename = NULL)
}
\arguments{
\item{beta}{Data frame, including all \code{samples} as columns.}

\item{samples}{Character, specifying sample names in \code{beta}.}

\item{main}{As in 'plot'.}

\item{xlab}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot the density of gene beta scores in two samples.
}
\note{
See the vignette for an example of density plot for beta score deviation.
Note that the source code of \code{DensityView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::DensityView}
or \code{getMethod("DensityView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/DensityView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
DensityView(dd, samples=c("D7_R1", "D7_R2", "PLX7_R1", "PLX7_R2"))
#or
DensityView(dd[, 3:6])

}
\seealso{
\code{\link{ViolinView}}
}
\author{
Wubing Zhang
}
