% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.DAVID.R
\docType{methods}
\name{enrich.DAVID}
\alias{enrich.DAVID}
\alias{enrichDAVID}
\title{Do enrichment analysis using DAVID}
\usage{
enrich.DAVID(gene, universe = NULL, david.user, idType = "ENTREZ_GENE_ID",
  minGSSize = 2, maxGSSize = 500, annotation = "GOTERM_BP_FAT",
  pvalueCutoff = 1, pAdjustMethod = "BH", qvalueCutoff = 0.2)
}
\arguments{
\item{gene}{Character vector, specifying the genelist to do enrichment analysis.}

\item{universe}{Character vector, specifying the backgound genelist, default is whole genome.}

\item{david.user}{Character, specifying a valid DAVID user account.}

\item{idType}{Character, indicating the gene id type of input genelist, such as "ENTREZ_GENE_ID"(default).}

\item{minGSSize}{Minimal size of each geneSet for testing.}

\item{maxGSSize}{Maximal size of each geneSet for analyzing.}

\item{annotation}{Geneset category for testing, GOTERM_BP_FAT(default).}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{qvalueCutoff}{Qvalue cutoff.}
}
\value{
A enrichResult instance.
}
\description{
an update version of DAVIDWebService to do enrichment analysis
}
\note{
This function depends on network and DAVID account, so don't show in the vignette.
The source can be found by typing \code{MAGeCKFlute:::enrich.DAVID}
or \code{getMethod("enrich.DAVID")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/enrich.DAVID.R}
Users should find it easy to customize this function.
}
\examples{
\dontrun{
 data(MLE_Data)
 universe = TransGeneID(MLE_Data$Gene, "SYMBOL", "ENTREZID", organism = "hsa")
 genes = universe[1:50]
	# Before running this example, you need to have a david account.
	enrichRes <- enrich.DAVID(genes, universe, david.user="david.user@edu.com")
	head(enrichRes@result)
}


}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.GOstats}}

\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
