% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GOstats.R
\docType{methods}
\name{enrich.GOstats}
\alias{enrich.GOstats}
\alias{enrichGOstats}
\title{Do enrichment analysis using GOstats}
\usage{
enrich.GOstats(gene, universe = NULL, type = c("KEGG", "BP", "MF", "CC"),
  organism = "hsa", pvalueCutoff = 1, pAdjustMethod = "BH")
}
\arguments{
\item{gene}{A character vector, specifying the genelist to do enrichment analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{type}{Geneset category for testing, KEGG(default).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using GOstats method
}
\note{
See the vignette for an example of enrichment analysis using GOstats.
The source can be found by typing \code{MAGeCKFlute:::enrich.GOstats}
or \code{getMethod("enrich.GOstats")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/enrich.GOstats.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
universe = TransGeneID(MLE_Data$Gene, "SYMBOL", "ENTREZID", organism = "hsa")
genes = universe[1:50]
enrichRes <- enrich.GOstats(genes, universe, type="BP")
head(enrichRes@result)

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
