% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{Gene set enrichment analysis}
\usage{
enrich.GSE(geneList, keytype = "Entrez", type = "CORUM+KEGG",
  organism = "hsa", pvalueCutoff = 0.25, limit = c(1, 120),
  gmtpath = NA)
}
\arguments{
\item{geneList}{A order ranked numeric vector with geneid as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'CORUM', 'CPX' (ComplexPortal),
'GOBP', 'GOMF', 'GOCC', 'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID',
or any combination of them (e.g. 'GOBP+GOMF+CORUM'), or 'All' (all categories).}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{limit}{A two-length vector (default: c(1, 120)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
A enrichResult instance.
}
\description{
A universal gene set enrichment analysis tools
}
\examples{
data(geneList, package = "DOSE")
\dontrun{
    enrichRes = enrich.GSE(geneList)
    head(slot(enrichRes, "result"))
}

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.ORT}}

\code{\link{EnrichAnalyzer}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
