% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{GSEA}
\usage{
enrich.GSE(geneList, type = "KEGG", organism = "hsa", minGSSize = 10,
  maxGSSize = 500, pvalueCutoff = 0.05, pAdjustMethod = "BH")
}
\arguments{
\item{geneList}{a order ranked numeric vector with geneid as names.}

\item{type}{a character, indicating geneset category for testing, "KEGG"(default).}

\item{organism}{a character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{minGSSize}{minimal size of each geneSet for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing.}

\item{pvalueCutoff}{pvalue cutoff.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}
}
\value{
A enrichResult instance.
}
\description{
A universal gene set enrichment analysis tools
}
\note{
See the vignette for an example of GSEA
The source can be found by typing \code{MAGeCKFlute:::enrich.GSE}
or \code{getMethod("enrich.GSE")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/enrich.GSE.R}
Users should find it easy to customize this function.
}
\examples{
 data(MLE_Data)
 universe = id2eg(MLE_Data$Gene, "SYMBOL")[,"ENTREZID"]
 geneList = MLE_Data$D7_R1.beta
 names(geneList) = universe
 geneList = geneList[!is.na(universe)]
	enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
	head(enrichRes@result)
	term = enrichRes@result$ID[1]
	gseaplot(enrichRes, term)


}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GOstats}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
