% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{GSEA}
\usage{
enrich.GSE(geneList, keytype = "Entrez", type = "CORUM",
  organism = "hsa", pvalueCutoff = 0.25, pAdjustMethod = "BH",
  limit = c(3, 50), gmtpath = NA)
}
\arguments{
\item{geneList}{A order ranked numeric vector with geneid as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'CORUM', 'CPX' (ComplexPortal),
'GOBP', 'GOMF', 'GOCC', 'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID',
or any combination of them (e.g. 'GOBP+GOMF+CORUM'), or 'All' (all categories).}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
A enrichResult instance.
}
\description{
A universal gene set enrichment analysis tools
}
\examples{
data(geneList, package = "DOSE")
\dontrun{
    enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
    head(slot(enrichRes, "result"))
}

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
