% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{Gene set enrichment analysis}
\usage{
enrich.GSE(
  geneList,
  keytype = "Symbol",
  type = "GOBP",
  organism = "hsa",
  pvalueCutoff = 0.25,
  limit = c(2, 200),
  gmtpath = NULL,
  by = "fgsea",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{geneList}{A order ranked numeric vector with geneid as names}

\item{keytype}{"Entrez", "Ensembl", or "Symbol"}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (KEGG, REACTOME, C2_CP), GO (GOBP, GOCC, GOMF),
MSIGDB (C1, C2 (C2_CP (C2_CP_PID, C2_CP_BIOCARTA), C2_CGP),
C3 (C3_MIR, C3_TFT), C4, C6, C7, HALLMARK)
and Complex (CORUM). Any combination of them are also accessible
(e.g. 'GOBP+GOMF+KEGG+REACTOME')}

\item{organism}{'hsa' or 'mmu'}

\item{pvalueCutoff}{FDR cutoff}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets for enrichent analysis}

\item{gmtpath}{The path to customized gmt file}

\item{by}{One of 'fgsea' or 'DOSE'}

\item{verbose}{Boolean}

\item{...}{Other parameter}
}
\value{
An enrichResult instance
}
\description{
A universal gene set enrichment analysis tools
}
\examples{
data(geneList, package = "DOSE")
\dontrun{
    enrichRes = enrich.GSE(geneList, keytype = "entrez")
    head(slot(enrichRes, "result"))
}

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.ORT}}

\code{\link{EnrichAnalyzer}}

\code{\link[clusterProfiler]{gseGO}}

\code{\link[clusterProfiler]{gseKEGG}}

\code{\link[clusterProfiler]{GSEA}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
