% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.GSE.R
\docType{methods}
\name{enrich.GSE}
\alias{enrich.GSE}
\alias{enrichGSE}
\title{GSEA}
\usage{
enrich.GSE(geneList, type = "MsigDB_c2_h", organism = "hsa",
  minGSSize = 10, maxGSSize = 500, pvalueCutoff = 0.25,
  pAdjustMethod = "BH")
}
\arguments{
\item{geneList}{A order ranked numeric vector with geneid as names.}

\item{type}{A character, indicating geneset category for testing, "MsigDB_c2_h"(default).}

\item{organism}{A character, specifying organism, only 'human' is available.}

\item{minGSSize}{Minimal size of each geneSet for testing.}

\item{maxGSSize}{Maximal size of each geneSet for analyzing.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}
}
\value{
A enrichResult instance.
}
\description{
A universal gene set enrichment analysis tools
}
\examples{
data(geneList, package = "DOSE")
enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
head(enrichRes@result)

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GOstats}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
