% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.HGT.R
\docType{methods}
\name{enrich.HGT}
\alias{enrich.HGT}
\alias{Hypergeometric}
\title{Do enrichment analysis using Hypergeometric test}
\usage{
enrich.HGT(geneList, keytype = "Entrez", type = "CORUM+KEGG",
  organism = "hsa", pvalueCutoff = 0.05, limit = c(1, 120),
  universe = NULL, gmtpath = NULL)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'GOBP', 'GOMF', 'GOCC', 'KEGG',
'BIOCARTA', 'REACTOME', 'CORUM', 'PID', 'HARKMARK', 'c2', 'c6', 'c7', or any
combination of them (e.g. 'GOBP+GOMF').}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{limit}{A two-length vector (default: c(1, 120)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using Hypergeometric test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:300]
enrichRes <- enrich.HGT(genes, type = "KEGG")
head(slot(enrichRes, "result"))

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{EnrichAnalyzer}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
