% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.HGT.R
\docType{methods}
\name{enrich.HGT}
\alias{enrich.HGT}
\alias{Hypergeometric}
\title{Do enrichment analysis using Hypergeometric test}
\usage{
enrich.HGT(geneList, keytype = "Symbol", type = "Pathway+GOBP",
  organism = "hsa", pvalueCutoff = 0.25, limit = c(2, 200),
  universe = NULL, gmtpath = NULL)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (PID, KEGG, REACTOME, BIOCARTA, C2CP), GO (GOBP, GOCC, GOMF),
Complex (CORUM, CPX), c1, c2, c3, c4, c6, c7, HALLMARK. It also allows any
combination of them (e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using Hypergeometric test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:300]
enrichRes <- enrich.HGT(genes, type = "KEGG", keytype = "entrez")
head(slot(enrichRes, "result"))

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{EnrichAnalyzer}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
