% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.HGT.R
\docType{methods}
\name{enrich.HGT}
\alias{enrich.HGT}
\alias{Hypergeometric}
\title{Do enrichment analysis using Hypergeometric test}
\usage{
enrich.HGT(geneList, universe = NULL, keytype = "Entrez",
  type = "GOBP+GOMF", organism = "hsa", pvalueCutoff = 0.25,
  pAdjustMethod = "BH", limit = c(3, 50), gmtpath = NA)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'GOBP+GOMF' (default), 'GOBP', 'GOMF', 'GOCC',
'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID', or 'All' and any combination of them,
such as 'KEGG+BIOCARTA+REACTOME+GOBP+GOCC+GOMF+EHMN+PID+WikiPathways'.}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{gmtpath}{The path to gmt file.}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using Hypergeometric test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
enrichRes <- enrich.HGT(genes, type = "KEGG")
head(enrichRes@result)

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
