% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.ORT.R
\docType{methods}
\name{enrich.ORT}
\alias{enrich.ORT}
\alias{enrichORT}
\title{Do enrichment analysis using over-representation test}
\usage{
enrich.ORT(geneList, keytype = "Entrez", type = "CORUM",
  organism = "hsa", pvalueCutoff = 0.25, pAdjustMethod = "BH",
  limit = c(3, 50), universe = NULL, gmtpath = NA)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, one of 'CORUM', 'CPX' (ComplexPortal),
'GOBP', 'GOMF', 'GOCC', 'KEGG', 'BIOCARTA', 'REACTOME', 'WikiPathways', 'EHMN', 'PID',
or any combination of them (e.g. 'GOBP+GOMF+CORUM'), or 'All' (all categories).}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A two-length vector (default: c(3, 50)), specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{gmtpath}{The path to customized gmt file.}
}
\value{
A enrichedResult instance.
}
\description{
Do enrichment analysis using over-representation test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
enrichedRes <- enrich.ORT(genes)
head(slot(enrichedRes, "result"))

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.GSE}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
