% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.ORT.R
\docType{methods}
\name{enrich.ORT}
\alias{enrich.ORT}
\alias{enrichORT}
\title{Do enrichment analysis using over-representation test}
\usage{
enrich.ORT(gene, universe = NULL, type = "KEGG", organism = "hsa",
  pvalueCutoff = 0.05, qvalueCutoff = 0.2, pAdjustMethod = "BH",
  minGSSize = 2, maxGSSize = 500)
}
\arguments{
\item{gene}{a character vector, specifying the genelist to do enrichment analysis.}

\item{universe}{a character vector, specifying the backgound genelist, default is whole genome.}

\item{type}{geneset category for testing, KEGG(default).}

\item{organism}{a character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{pvalue cutoff.}

\item{qvalueCutoff}{qvalue cutoff.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{minGSSize}{minimal size of each geneSet for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing.}
}
\value{
A enrichResult instance.
}
\description{
Do enrichment analysis using over-representation test
}
\note{
See the vignette for an example of enrichment analysis using over-representation test
The source can be found by typing \code{MAGeCKFlute:::enrich.ORT}
or \code{getMethod("enrich.ORT")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/enrich.ORT.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
universe = TransGeneID(MLE_Data$Gene, "SYMBOL", "ENTREZID", organism = "hsa")
genes = TransGeneID(Core_Essential[1:200], "SYMBOL", "ENTREZID", organism = "hsa")
enrichRes <- enrich.ORT(genes, universe)
head(enrichRes@result)


}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GOstats}}

\code{\link{enrich.GSE}}

\code{\link{enrichment_analysis}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
