% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.ORT.R
\docType{methods}
\name{enrich.ORT}
\alias{enrich.ORT}
\alias{enrichORT}
\title{Do enrichment analysis using over-representation test}
\usage{
enrich.ORT(
  geneList,
  keytype = "Symbol",
  type = "GOBP",
  organism = "hsa",
  pvalueCutoff = 0.25,
  limit = c(2, 200),
  universe = NULL,
  gmtpath = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (KEGG, REACTOME, C2_CP), GO (GOBP, GOCC, GOMF),
MSIGDB (C1, C2 (C2_CP (C2_CP_PID, C2_CP_BIOCARTA), C2_CGP),
C3 (C3_MIR, C3_TFT), C4, C6, C7, HALLMARK)
and Complex (CORUM). Any combination of them are also accessible
(e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{organism}{'hsa' or 'mmu'.}

\item{pvalueCutoff}{FDR cutoff.}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{gmtpath}{The path to customized gmt file.}

\item{verbose}{Boolean}

\item{...}{Other parameter}
}
\value{
An enrichedResult instance.
}
\description{
Do enrichment analysis using over-representation test
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
enrichedRes <- enrich.ORT(genes, keytype = "entrez")
head(slot(enrichedRes, "result"))

}
\seealso{
\code{\link{enrich.HGT}}

\code{\link{enrich.GSE}}

\code{\link{EnrichAnalyzer}}

\code{\link[clusterProfiler]{enrichGO}}

\code{\link[clusterProfiler]{enrichKEGG}}

\code{\link[clusterProfiler]{enricher}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Wubing Zhang
}
