% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichAB.R
\docType{methods}
\name{EnrichAB}
\alias{EnrichAB}
\title{Enrichment analysis for Positive and Negative selection genes}
\usage{
EnrichAB(data, pvalue = 0.25, enrich_method = "ORT", organism = "hsa",
  adjust = "BH", filename = NULL, out.dir = ".", gsea = FALSE,
  width = 6.5, height = 4, ...)
}
\arguments{
\item{data}{A data frame containing columns "diff", with rownames of Entrez IDs.}

\item{pvalue}{Pvalue cutoff.}

\item{enrich_method}{One of "ORT"(Over-Representing Test), "DAVID", "GOstats", and "HGT"(HyperGemetric test).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{adjust}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{filename}{Suffix of output file name. NULL(default) means no output.}

\item{out.dir}{Path to save plot to (combined with filename).}

\item{gsea}{Boolean, specifying if do GSEA for GroupA and GroupB genes. Default gsea=FALSE.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
A list containing enrichment results for each group genes. This list contains items four
items, \code{keggA}, \code{keggB}, \code{bpA}, \code{bpB}. Four items are all list object, containing
subitems of \code{gridPlot} and \code{enrichRes}. \code{gridPlot} is a ggplot object, and
\code{enrichRes} is a enrichResult instance
}
\description{
Do enrichment analysis for selected genes, in which positive selection and negative selection
are termed as GroupA and GroupB
}
\seealso{
\code{\link{EnrichSquare}}


data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
}
\author{
Binbin Wang
}
