% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichAB.R
\docType{methods}
\name{enrichAB}
\alias{enrichAB}
\alias{EnrichAB}
\title{Enrichment analysis for Positive and Negative selection genes}
\usage{
EnrichAB(data, pvalue = 0.05, enrich_method = "ORT", organism = "hsa",
  adjust = "BH", filename = NULL, out.dir = ".", gsea = FALSE)
}
\arguments{
\item{data}{a data frame containing columns of "ENTREZID" and "diff".}

\item{pvalue}{pvalue cutoff.}

\item{enrich_method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), "DAVID",
"GOstats", and "HGT"(HyperGemetric test), or index from 1 to 5}

\item{organism}{a character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{adjust}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{filename}{suffix of output file name. NULL(default) means no output.}

\item{out.dir}{Path to save plot to (combined with filename).}

\item{gsea}{boolean, specifying if do GSEA for GroupA and GroupB genes. Default gsea=FALSE.}
}
\value{
a list containing enrichment results for each group genes. This list contains items four
items, \code{keggA}, \code{keggB}, \code{bpA}, \code{bpB}. Four items are all list object, containing
subitems of \code{gridPlot} and \code{enrichRes}. \code{gridPlot} is a ggplot object, and
\code{enrichRes} is a enrichResult instance
}
\description{
Do enrichment analysis for selected genes, in which positive selection and negative selection
are termed as GroupA and GroupB
}
\note{
See the vignette for an example of EnrichAB
The source can be found by typing \code{MAGeCKFlute:::EnrichAB}
or \code{getMethod("EnrichAB")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/EnrichAB.R}
Users should find it easy to customize this function.
}
\seealso{
\code{\link{EnrichSquare}}


data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
}
\author{
Binbin Wang
}
