% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedGeneView.R
\docType{methods}
\name{EnrichedGeneView}
\alias{EnrichedGeneView}
\title{Visualize selected genes in enriched genesets}
\usage{
EnrichedGeneView(enrichment, geneList, keytype = "Symbol",
  gene_cutoff = c(-log2(1.5), log2(1.5)), top = 5, bottom = 5,
  charLength = 40, filename = NULL, width = 7, height = 5, ...)
}
\arguments{
\item{enrichment}{A data frame of enrichment result.}

\item{geneList}{The geneList used in enrichment analysis.}

\item{keytype}{"Entrez" or "Symbol".}

\item{gene_cutoff}{A tow-length numeric vector, specifying cutoff for negative and positive selections.}

\item{top}{An integer, specifying the number of top enriched terms to show.}

\item{bottom}{An integer, specifying the number of bottom enriched terms to show.}

\item{charLength}{Integer, specifying max length of enriched term name to show as coordinate lab.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Visualize selected genes in enriched genesets
}
\examples{
data(geneList, package = "DOSE")
enrichRes <- enrich.GSE(geneList)
EnrichedGeneView(enrichment=as.data.frame(enrichRes), geneList, keytype = "Entrez")
}
\author{
Wubing Zhang
}
