% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedView.R
\docType{methods}
\name{EnrichedGSEView}
\alias{EnrichedGSEView}
\alias{enrichgseview}
\title{View enriched terms in GSEA}
\usage{
EnrichedGSEView(enrichment, plotTitle = NULL, termNum = 15,
  charLength = 40, filename = NULL, width = 5, height = 4, ...)
}
\arguments{
\item{enrichment}{A data frame of enrichment result, with columns of ID, Description, p.adjust and NES}

\item{plotTitle}{Same as 'title' in 'plot'.}

\item{termNum}{Integer, specifying number of top enriched terms to show}

\item{charLength}{Integer, specifying max length of enriched term name to show as coordinate lab}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Grid plot for enriched terms in GSEA
}
\examples{
\dontrun{
    data(geneList, package = "DOSE")
    enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
    EnrichedGSEView(enrichRes@result, plotTitle = "GSEA Analysis")
}

}
\seealso{
\code{\link{EnrichedView}}
}
\author{
Wubing Zhang
}
