% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedView.R
\docType{methods}
\name{EnrichedGSEView}
\alias{EnrichedGSEView}
\alias{enrichgseview}
\title{View enriched terms in GSEA}
\usage{
EnrichedGSEView(enrichment, plotTitle = NULL, gridColour = "blue",
  termNum = 10, charLength = 50)
}
\arguments{
\item{enrichment}{a data frame of enrichment result, with columns of ID, Description, p.adjust and NES}

\item{plotTitle}{same as 'title' in 'plot'.}

\item{gridColour}{color of grids.}

\item{termNum}{integer, specifying number of top enriched terms to show}

\item{charLength}{integer, specifying max length of enriched term name to show as coordinate lab}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Grid plot for enriched terms in GSEA
}
\note{
See the vignette for an example of EnrichedGSEView
The source can be found by typing \code{MAGeCKFlute:::EnrichedGSEView}
or \code{getMethod("EnrichedGSEView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/EnrichedGSEView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
universe = id2eg(MLE_Data$Gene, "SYMBOL")[,"ENTREZID"]
geneList = MLE_Data$D7_R1.beta
names(geneList) = universe
geneList = geneList[!is.na(universe)]
enrichRes = enrich.GSE(geneList, type = "KEGG", organism="hsa")
EnrichedGSEView(enrichRes@result, plotTitle = "GSEA Analysis")



}
\seealso{
\code{\link{EnrichedView}}
}
\author{
Wubing Zhang
}
