% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedView.R
\docType{methods}
\name{EnrichedView}
\alias{EnrichedView}
\alias{enrichview}
\title{View enriched terms}
\usage{
EnrichedView(enrichment, plotTitle = NULL, color = "#3f90f7",
  termNum = 15, charLength = 40, filename = NULL, width = 5,
  height = 4, ...)
}
\arguments{
\item{enrichment}{A data frame of enrichment result, with columns of ID, Description, p.adjust and Count.}

\item{plotTitle}{Same as 'title' in 'plot'.}

\item{color}{Color of nodes.}

\item{termNum}{Integer, specifying number of top enriched terms to show.}

\item{charLength}{Integer, specifying max length of enriched term name to show as coordinate lab.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Grid plot for enriched terms
}
\note{
See the vignette for an example of EnrichedView
The source can be found by typing \code{MAGeCKFlute:::EnrichedView}
or \code{getMethod("EnrichedView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/EnrichedView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
universe = TransGeneID(MLE_Data$Gene, "SYMBOL", "ENTREZID", organism = "hsa")
genes = universe[1:50]
enrichRes <- enrich.HGT(genes, universe)
EnrichedView(enrichment=enrichRes@result, plotTitle = "Hypergemetric test")


}
\seealso{
\code{\link{KeggPathwayView}}

\code{\link{EnrichedGSEView}}
}
\author{
Feizhen Wu
}
