% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichedView.R
\docType{methods}
\name{EnrichedView}
\alias{EnrichedView}
\alias{enrichgseview}
\title{View enriched terms}
\usage{
EnrichedView(enrichment, rank_by = "p.adjust", top = 5, bottom = 5,
  custom_pid = NULL, x = "LogP", charLength = 40, filename = NULL,
  width = 7, height = 4, ...)
}
\arguments{
\item{enrichment}{A data frame of enrichment result, with columns of ID, Description, p.adjust and NES.}

\item{rank_by}{"p.adjust" or "NES", specifying the indices for ranking pathways.}

\item{top}{An integer, specifying the number of top enriched terms to show.}

\item{bottom}{An integer, specifying the number of bottom enriched terms to show.}

\item{custom_pid}{A character vector (pathway IDs), customizing pathways to show.}

\item{x}{Character, "NES" or "LogP", indicating the variable on the x-axis.}

\item{charLength}{Integer, specifying max length of enriched term name to show as coordinate lab.}

\item{filename}{Figure file name to create on disk. Default filename="NULL".}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Grid plot for enriched terms
}
\examples{
\dontrun{
    data(geneList, package = "DOSE")
    enrichRes = enrich.GSE(geneList, organism="hsa")
    EnrichedView(slot(enrichRes, "result"))
}
}
\seealso{
\code{\link{EnrichedView}}
}
\author{
Wubing Zhang
}
