% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\docType{methods}
\name{enrichment_analysis}
\alias{enrichment_analysis}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
enrichment_analysis(geneList, universe = NULL, method = 1, type = "KEGG",
  organism = "hsa", pvalueCutoff = 0.05, qvalueCutoff = 1,
  pAdjustMethod = "BH", minGSSize = 2, maxGSSize = 500,
  plotTitle = NULL, gridColour = "blue")
}
\arguments{
\item{geneList}{a character vector or a ranked numeric vector(for GSEA) with names of geneid,
specifying the genelist to do enrichment analysis.}

\item{universe}{a character vector, specifying the backgound genelist, default is whole genome.}

\item{method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), "DAVID",
"GOstats", and "HGT"(HyperGemetric test), or index from 1 to 5}

\item{type}{geneset category for testing, KEGG(default).}

\item{organism}{a character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{pvalue cutoff.}

\item{qvalueCutoff}{qvalue cutoff.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{minGSSize}{minimal size of each geneSet for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing.}

\item{plotTitle}{same as 'title' in 'plot'.}

\item{gridColour}{color of grids.}
}
\value{
a list, including two items, \code{gridPlot} and \code{enrichRes}. \code{gridPlot} is
a ggplot object, and \code{enrichRes} is a enrichResult instance.
}
\description{
Enrichment analysis
}
\note{
See the vignette for an example of enrichment analysis
The source can be found by typing \code{MAGeCKFlute:::enrichment_analysis}
or \code{getMethod("enrichment_analysis")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/enrichment_analysis.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
universe = TransGeneID(MLE_Data$Gene, "SYMBOL", "ENTREZID", organism = "hsa")
genes = TransGeneID(Core_Essential[1:200], "SYMBOL", "ENTREZID", organism = "hsa")
keggA = enrichment_analysis(geneList = genes, universe=universe,
                         method = "HGT",type = "KEGG",
                         organism = "hsa", gridColour="#e41a1c")
print(keggA$gridPlot)


}
\seealso{
\code{\link{enrich.GOstats}}

\code{\link{enrich.DAVID}}

\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrich.HGT}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Feizhen Wu
}
