% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\docType{methods}
\name{enrichment_analysis}
\alias{enrichment_analysis}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
enrichment_analysis(geneList, universe = NULL, method = "ORT",
  keytype = "Entrez", type = "KEGG", organism = "hsa",
  pvalueCutoff = 0.25, pAdjustMethod = "BH", limit = c(3, 50),
  plotTitle = NULL, color = "#3f90f7")
}
\arguments{
\item{geneList}{A numeric vector with gene as names.}

\item{universe}{A character vector, specifying the backgound genelist, default is whole genome.}

\item{method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), and "HGT"(HyperGemetric test).}

\item{keytype}{"Entrez" or "Symbol".}

\item{type}{Geneset category for testing, "MsigDB" (default), "KEGG", "BP", "CC", "MF", "DO",
"MKEGG", "NCG" or "Others" (`gmtpath` is required).}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{pvalueCutoff}{Pvalue cutoff.}

\item{pAdjustMethod}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{limit}{A integer vector of length two, specifying the limit of geneset size.}

\item{plotTitle}{Same as 'title' in 'plot'.}

\item{color}{Color of points.}
}
\value{
A list, including two items, \code{gridPlot} and \code{enrichRes}. \code{gridPlot} is
a ggplot object, and \code{enrichRes} is a enrichResult instance.
}
\description{
Enrichment analysis
}
\examples{
data(geneList, package = "DOSE")
genes <- geneList[1:100]
keggA = enrichment_analysis(genes, method = "HGT", type = "KEGG")
print(keggA$gridPlot)

}
\seealso{
\code{\link{enrich.GSE}}

\code{\link{enrich.ORT}}

\code{\link{enrich.HGT}}

\code{\link[DOSE]{enrichResult-class}}
}
\author{
Feizhen Wu
}
