% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichSquare.R
\docType{methods}
\name{EnrichSquare}
\alias{EnrichSquare}
\title{Enrichment analysis for selected treatment related genes}
\usage{
EnrichSquare(
  beta,
  id = "Gene",
  keytype = "Symbol",
  x = "Control",
  y = "Treatment",
  enrich_method = "HGT",
  top = 5,
  organism = "hsa",
  limit = c(2, 100),
  filename = NULL,
  out.dir = ".",
  width = 6.5,
  height = 4,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{beta}{Data frame, with columns of "Gene", "group", and "Diff".}

\item{id}{A character, indicating the gene column in the data.}

\item{keytype}{A character, "Symbol" or "Entrez".}

\item{x}{A character, indicating the x-axis in the 9-square scatter plot.}

\item{y}{A character, indicating the y-axis in the 9-square scatter plot.}

\item{enrich_method}{One of "ORT"(Over-Representing Test) and "HGT"(HyperGemetric test).}

\item{top}{An integer, specifying the number of pathways to show.}

\item{organism}{"hsa" or "mmu".}

\item{limit}{A two-length vector, specifying the min and
max size of pathways for enrichent analysis.}

\item{filename}{Suffix of output file name. NULL(default) means no output.}

\item{out.dir}{Path to save plot to (combined with filename).}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{verbose}{Boolean.}

\item{...}{Other available parameters in ggsave.}
}
\value{
A list containing enrichment results for each group genes.
Each item in the returned list has two sub items:
\item{gridPlot}{an object created by \code{ggplot}, which can be assigned and further customized.}
\item{enrichRes}{a enrichResult instance.}
}
\description{
Do enrichment analysis for selected treatment related genes in 9-squares
}
\author{
Wubing Zhang
}
