% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichSquare.R
\docType{methods}
\name{EnrichSquare}
\alias{EnrichSquare}
\title{Enrichment analysis for selected treatment related genes}
\usage{
EnrichSquare(beta, pvalue = 1, enrich_method = "ORT", organism = "hsa",
  adjust = "BH", filename = NULL, out.dir = ".")
}
\arguments{
\item{beta}{Data frame, which has columns of 'ENTREZID' and 'group'.}

\item{pvalue}{Pvalue cutoff.}

\item{enrich_method}{One of "ORT"(Over-Representing Test), "GSEA"(Gene Set Enrichment Analysis), "DAVID",
"GOstats", and "HGT"(HyperGemetric test), or index from 1 to 5}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default), and "mmu" or "Mouse"}

\item{adjust}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{filename}{Suffix of output file name. NULL(default) means no output.}

\item{out.dir}{Path to save plot to (combined with filename).}
}
\value{
A list containing enrichment results for each group genes. This list contains several elements:
\item{kegg1}{a list record enriched KEGG pathways for Group1 genes in 9-Square}
\item{kegg2}{a list record enriched KEGG pathways for Group2 genes in 9-Square}
\item{kegg3}{a list record enriched KEGG pathways for Group3 genes in 9-Square}
\item{kegg4}{a list record enriched KEGG pathways for Group4 genes in 9-Square}
\item{kegg13}{a list record enriched KEGG pathways for Group1&Group3 genes in 9-Square}
\item{kegg14}{a list record enriched KEGG pathways for Group1&Group4 genes in 9-Square}
\item{kegg23}{a list record enriched KEGG pathways for Group2&Group3 genes in 9-Square}
\item{kegg24}{a list record enriched KEGG pathways for Group2&Group4 genes in 9-Square}
\item{kegg1234}{a list record enriched KEGG pathways for Group1&Group2&Group3&Group4 genes in 9-Square}
\item{bp1}{a list record enriched GO BP terms for Group1 genes in 9-Square}
\item{bp2}{a list record enriched GO BP terms for Group2 genes in 9-Square}
\item{bp3}{a list record enriched GO BP terms for Group3 genes in 9-Square}
\item{bp4}{a list record enriched GO BP terms for Group4 genes in 9-Square}
\item{bp13}{a list record enriched GO BP terms for Group1&Group3 genes in 9-Square}
\item{bp14}{a list record enriched GO BP terms for Group1&Group4 genes in 9-Square}
\item{bp23}{a list record enriched GO BP terms for Group2&Group3 genes in 9-Square}
\item{bp24}{a list record enriched GO BP terms for Group2&Group4 genes in 9-Square}
\item{bp1234}{a list record enriched GO BP terms for Group1&Group2&Group3&Group4 genes in 9-Square}
Each item in the returned list has two sub items:
\item{gridPlot}{an object created by \code{ggplot}, which can be assigned and further customized.}
\item{enrichRes}{a enrichResult instance.}
}
\description{
Do enrichment analysis for selected treatment related genes in 9-squares
}
\note{
See the vignette for an example of EnrichSquare
The source can be found by typing \code{MAGeCKFlute:::EnrichSquare}
or \code{getMethod("EnrichSquare")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/EnrichSquare.R}
Users should find it easy to customize this function.
}
\seealso{
\code{\link{SquareView}}

\code{\link{EnrichSquare}}

 # Read beta score from gene summary table in MAGeCK MLE results
}
\author{
Wubing Zhang
}
