% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichSquare.R
\docType{methods}
\name{EnrichSquare}
\alias{EnrichSquare}
\title{Enrichment analysis for selected treatment related genes}
\usage{
EnrichSquare(beta, pvalue = 0.05, enrich_method = "ORT",
  organism = "hsa", limit = c(1, 120), filename = NULL,
  out.dir = ".", width = 6.5, height = 4, ...)
}
\arguments{
\item{beta}{Data frame, with rownames of Entrez IDs, which contains columns of 'group' and 'diff'.}

\item{pvalue}{Pvalue cutoff.}

\item{enrich_method}{One of "ORT"(Over-Representing Test) and "HGT"(HyperGemetric test).}

\item{organism}{"hsa" or "mmu".}

\item{limit}{A two-length vector (default: c(1, 120)), specifying the min and
max size of pathways for enrichent analysis.}

\item{filename}{Suffix of output file name. NULL(default) means no output.}

\item{out.dir}{Path to save plot to (combined with filename).}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
A list containing enrichment results for each group genes. This list contains several elements:
\item{kegg1}{a list record enriched KEGG pathways for Group1 genes in 9-Square}
\item{kegg2}{a list record enriched KEGG pathways for Group2 genes in 9-Square}
\item{kegg3}{a list record enriched KEGG pathways for Group3 genes in 9-Square}
\item{kegg4}{a list record enriched KEGG pathways for Group4 genes in 9-Square}
\item{kegg12}{a list record enriched KEGG pathways for Group1&Group2 genes in 9-Square}
\item{kegg13}{a list record enriched KEGG pathways for Group1&Group3 genes in 9-Square}
\item{kegg24}{a list record enriched KEGG pathways for Group2&Group4 genes in 9-Square}
\item{kegg34}{a list record enriched KEGG pathways for Group3&Group4 genes in 9-Square}
\item{go1}{a list record enriched GO terms for Group1 genes in 9-Square}
\item{go2}{a list record enriched GO terms for Group2 genes in 9-Square}
\item{go3}{a list record enriched GO terms for Group3 genes in 9-Square}
\item{go4}{a list record enriched GO terms for Group4 genes in 9-Square}
\item{go12}{a list record enriched GO terms for Group1&Group2 genes in 9-Square}
\item{go13}{a list record enriched GO terms for Group1&Group3 genes in 9-Square}
\item{go24}{a list record enriched GO terms for Group2&Group4 genes in 9-Square}
\item{go34}{a list record enriched GO terms for Group3&Group4 genes in 9-Square}

Each item in the returned list has two sub items:
\item{gridPlot}{an object created by \code{ggplot}, which can be assigned and further customized.}
\item{enrichRes}{a enrichResult instance.}
}
\description{
Do enrichment analysis for selected treatment related genes in 9-squares
}
\author{
Wubing Zhang
}
