% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluteMLE.R
\docType{methods}
\name{FluteMLE}
\alias{FluteMLE}
\alias{flutemle}
\title{Downstream analysis based on MAGeCK-MLE result}
\usage{
FluteMLE(gene_summary, ctrlname, treatname, keytype = "Symbol",
  organism = "hsa", norm_method = "none", posControls = NULL,
  top = 10, bottom = 10, toplabels = NA, scale_cutoff = 2,
  limit = c(0, 200), pvalueCutoff = 0.25, enrich_method = "ORT",
  proj = NA, width = 10, height = 7, outdir = ".",
  view_allpath = FALSE)
}
\arguments{
\item{gene_summary}{A data frame or a file path to gene summary file generated by MAGeCK-MLE.}

\item{ctrlname}{A character vector, specifying the names of control samples.}

\item{treatname}{A character vector, specifying the names of treatment samples.}

\item{keytype}{"Entrez" or "Symbol".}

\item{organism}{"hsa" or "mmu".}

\item{norm_method}{One of "none" (default), "cell_cycle" or "loess".}

\item{top}{An integer, specifying number of top selected genes to be labeled in rank figure.}

\item{bottom}{An integer, specifying number of bottom selected genes to be labeled in rank figure.}

\item{toplabels}{A character vector, specifying interested genes to be labeled in rank figure.}

\item{scale_cutoff}{Boolean or numeric, specifying how many standard deviation will be used as cutoff.}

\item{limit}{A two-length vector, specifying the minimal and maximal size of gene sets for enrichent analysis.}

\item{pvalueCutoff}{A numeric, specifying pvalue cutoff of enrichment analysis, default 1.}

\item{enrich_method}{One of "ORT"(Over-Representing Test), "HGT"(HyperGemetric test) and
"GSEA"(Gene Set Enrichment Analysis).}

\item{proj}{A character, indicating the prefix of output file name, which can't contain special characters.}

\item{width}{The width of summary pdf in inches.}

\item{height}{The height of summary pdf in inches.}

\item{outdir}{Output directory on disk.}

\item{view_allpath}{Boolean, whether output all pathway view figures (time-consuming).}

\item{posControl}{A character vector, specifying a list of positive control gene symbols.}
}
\value{
All of the pipeline results is output into the \code{out.dir}/MAGeCKFlute_\code{proj},
which includes a pdf file and many folders. The pdf file 'FluteMLE_\code{proj}_\code{norm_method}.pdf' is the
summary of pipeline results. For each section in this pipeline, figures and useful data are
outputed to corresponding subfolders.
\itemize{
  \item {QC}: {Quality control}
  \item {Selection}: {Positive selection and negative selection.}
  \item {Enrichment}: {Enrichment analysis for positive and negative selection genes.}
  \item {PathwayView}: {Pathway view for top enriched pathways.}
}
}
\description{
Integrative analysis pipeline using the gene summary table in MAGeCK MLE results
}
\details{
MAGeCK-MLE can be used to analyze screen data from multi-conditioned experiments. MAGeCK-MLE
also normalizes the data across multiple samples, making them comparable to each other. The most important
ouput of MAGeCK MLE is `gene_summary` file, which includes the beta scores of multiple conditions and
the associated statistics. The `beta score`  for each gene describes how the gene is selected: a positive
beta score indicates a positive selection, and a negative beta score indicates a negative selection.

The downstream analysis includes identifying essential, non-essential, and target-associated genes,
and performing biological functional category analysis and pathway enrichment analysis of these genes.
The function also visualizes genes in the context of pathways to benefit users exploring screening data.
}
\examples{
data(mle.gene_summary)
\dontrun{
  # functional analysis for MAGeCK MLE results
  FluteMLE(mle.gene_summary, ctrlname = "dmso", treatname = "plx",
           proj = "PLX", pvalueCutoff = 0.25, organism = "hsa")
}

}
\seealso{
\code{\link{FluteRRA}}
}
\author{
Wubing Zhang
}
