% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluteMLE.R
\docType{methods}
\name{FluteMLE}
\alias{FluteMLE}
\alias{MLEpipeline,}
\alias{MAGeCKFlute-method}
\title{Downstream analysis based on MAGeCK-MLE result}
\usage{
FluteMLE(gene_summary, ctrlname = "Control", treatname = "Treatment",
  organism = "hsa", prefix = "Test", top = 10, bottom = 10,
  interestGenes = c(), pvalueCutoff = 1, adjust = "BH",
  enrich_kegg = "ORT", gsea = FALSE, posControl = NULL,
  scale_cutoff = 1, loess = FALSE, view_allpath = FALSE, outdir = ".")
}
\arguments{
\item{gene_summary}{Either a file path or a data frame, which contains columns of 'Gene',
\code{ctrlname}.beta and \code{treatname}.beta which corresponding to the parameter ctrlname and treatmname.}

\item{ctrlname}{A character vector, specifying the name of control samples.}

\item{treatname}{A character vector, specifying the name of treatment samples.}

\item{organism}{A character, specifying organism, such as "hsa" or "Human"(default),
and "mmu" or "Mouse".}

\item{prefix}{A character, indicating the prefix of output file name, which can't contain special characters.}

\item{top}{An integer, specifying number of top selected genes labeled in rank figure.}

\item{bottom}{An integer, specifying number of bottom selected genes labeled in rank figure.}

\item{interestGenes}{A character vector, specifying interested genes labeled in rank figure.}

\item{pvalueCutoff}{A numeric, specifying pvalue cutoff of enrichment analysis, default 1.}

\item{adjust}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{enrich_kegg}{One of "ORT"(Over-Representing Test), "DAVID", "GOstats", and "HGT"(HyperGemetric test),
specifying enrichment method used for kegg enrichment analysis.}

\item{gsea}{Boolean, indicating whether GSEA analysis is needed for positive and
negative selection genes.}

\item{posControl}{A file path or a character vector, specifying a list of gene entrezid as positive controls
used for cell cycle normalization.}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{loess}{Boolean, whether include loess normalization in the pipeline.}

\item{view_allpath}{Boolean, whether output all pathway view figures.}

\item{outdir}{Output directory on disk.}
}
\value{
All of the pipeline results is output into the \code{out.dir}/\code{prefix}_Results,
which includes a pdf file and many folders. The pdf file '{prefix}_Pipeline_results.pdf' is the
summary of pipeline results. For each section in this pipeline, figures and useful data are
outputed to corresponding subfolders.
Distribution_of_BetaScores: Density plot and violin plot of beta scores.
MAplot: Maplot for each normalized data.
Linear_Fitting_of_BetaScores: Linear fitting of beta scores indicates the difference of cell cycle
time between Control and Treatment samples.
Scatter_Treat_Ctrl: Positive selection and negative selection
Enrichment_Treat-Ctrl: Enrichment analysis for positive and negative selection genes
Pathview_Treat_Ctrl: Pathway view for top enriched pathways
Scatter_9Square: Using 9 Square to select drug related genes
Enrichment_9Square: Enrichment analysis for selected genes
Pathview_9Square: Pathway view for top enriched pathways
}
\description{
Integrative analysis pipeline using the gene summary table in MAGeCK MLE results
}
\details{
MAGeCK-MLE can be used to analyze screen data from multi-conditioned experiments. MAGeCK-MLE
also normalizes the data across multiple samples, making them comparable to each other. The most important
ouput of MAGeCK MLE is `gene_summary` file, which includes the beta scores of multiple conditions and
the associated statistics. The `beta score`  for each gene describes how the gene is selected: a positive
beta score indicates a positive selection, and a negative beta score indicates a negative selection.

The downstream analysis includes identifying essential, non-essential, and target-associated genes,
and performing biological functional category analysis and pathway enrichment analysis of these genes.
The function also visualizes genes in the context of pathways to benefit users exploring screening data.
}
\note{
See the vignette for an example of FluteMLE
The source can be found by typing \code{MAGeCKFlute:::FluteMLE}
or \code{getMethod("FluteMLE")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/FluteMLE.R}
Users should find it easy to customize this function.
}
\examples{
\dontrun{
  data(MLE_Data)
  # functional analysis for MAGeCK MLE results
  FluteMLE(MLE_Data, ctrlname=c("D7_R1","D7_R2"), treatname=c("PLX7_R1","PLX7_R2"),
           prefix="BRAF_D7", pvalueCutoff=0.05, organism="hsa")
}
}
\seealso{
\code{\link{FluteRRA}}
}
\author{
Wubing Zhang
}
