% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluteRRA.R
\docType{methods}
\name{FluteRRA}
\alias{FluteRRA}
\alias{RRApipeline}
\title{Downstream analysis based on MAGeCK-RRA result}
\usage{
FluteRRA(
  gene_summary,
  sgrna_summary = NULL,
  keytype = "Symbol",
  organism = "hsa",
  incorporateDepmap = TRUE,
  cell_lines = NA,
  lineages = "All",
  omitEssential = FALSE,
  top = 5,
  toplabels = NULL,
  scale_cutoff = 2,
  limit = c(2, 200),
  pvalueCutoff = 0.25,
  proj = NA,
  width = 12,
  height = 6,
  outdir = ".",
  verbose = TRUE
)
}
\arguments{
\item{gene_summary}{A file path or a data frame of gene summary data.}

\item{sgrna_summary}{A file path or a data frame of sgRNA summary data.}

\item{keytype}{"Entrez" or "Symbol".}

\item{organism}{"hsa" or "mmu".}

\item{incorporateDepmap}{Boolean, indicating whether incorporate Depmap data into analysis.}

\item{cell_lines}{A character vector, specifying the cell lines in Depmap to be considered.}

\item{lineages}{A character vector, specifying the lineages in Depmap to be considered.}

\item{omitEssential}{Boolean, indicating whether omit common essential genes from the downstream analysis.}

\item{top}{An integer, specifying number of top selected genes to be labeled in rank figure.}

\item{toplabels}{A character vector, specifying interested genes to be labeled in rank figure.}

\item{scale_cutoff}{Boolean or numeric, specifying how many standard deviation will be used as cutoff.}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{pvalueCutoff}{A numeric, specifying pvalue cutoff of enrichment analysis, default 1.}

\item{proj}{A character, indicating the prefix of output file name.}

\item{width}{The width of summary pdf in inches.}

\item{height}{The height of summary pdf in inches.}

\item{outdir}{Output directory on disk.}

\item{verbose}{Boolean}
}
\value{
All of the pipeline results is output into the \code{out.dir}/\code{proj}_Results,
which includes a pdf file and a folder named 'RRA'.
}
\description{
Integrative analysis pipeline using the gene summary table in MAGeCK RRA results
}
\details{
MAGeCK RRA allows for the comparison between two experimental conditions. It can identify
genes and sgRNAs are significantly selected between the two conditions. The most important output
of MAGeCK RRA is the file `gene_summary.txt`. MAGeCK RRA will output both the negative score and
positive score for each gene. A smaller score indicates higher gene importance. MAGeCK RRA  will
also output the statistical value for the scores of each gene. Genes that are significantly positively
and negatively selected can be identified based on the p-value or FDR.

The downstream analysis of this function includes identifying positive and negative selection genes,
and performing biological functional category analysis and pathway enrichment analysis of these genes.
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
file2 = file.path(system.file("extdata", package = "MAGeCKFlute"),
                  "testdata/rra.sgrna_summary.txt")
\dontrun{
    # Run the FluteRRA pipeline
    FluteRRA(file1, file2, proj="PLX", organism="hsa", incorporateDepmap = FALSE,
    scale_cutoff = 1, outdir = "./")
}

}
\seealso{
\code{\link{FluteMLE}}
}
\author{
Wubing Zhang
}
