% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluteRRA.R
\docType{methods}
\name{FluteRRA}
\alias{FluteRRA}
\alias{RRApipeline}
\title{Downstream analysis based on MAGeCK-RRA result}
\usage{
FluteRRA(gene_summary, sgrna_summary, lfcCutoff = c(-1, 1),
  organism = "hsa", limit = c(2, 200), pvalueCutoff = 0.25,
  prefix = "Test", width = 12, height = 6, outdir = ".")
}
\arguments{
\item{gene_summary}{A file path or a data frame of gene summary data.}

\item{sgrna_summary}{A file path or a data frame of sgRNA summary data.}

\item{lfcCutoff}{A two-length vector (default: c(-1, 1)), specifying the logFC cutoff
for negative selection and positive selection.}

\item{organism}{"hsa" or "mmu".}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets for enrichent analysis.}

\item{pvalueCutoff}{A numeric, specifying pvalue cutoff of enrichment analysis, default 1.}

\item{prefix}{A character, indicating the prefix of output file name.}

\item{width}{The width of summary pdf in inches.}

\item{height}{The height of summary pdf in inches.}

\item{outdir}{Output directory on disk.}
}
\value{
All of the pipeline results is output into the \code{out.dir}/\code{prefix}_Results,
which includes a pdf file and a folder named 'RRA'.
}
\description{
Integrative analysis pipeline using the gene summary table in MAGeCK RRA results
}
\details{
MAGeCK RRA allows for the comparison between two experimental conditions. It can identify
genes and sgRNAs are significantly selected between the two conditions. The most important output
of MAGeCK RRA is the file `gene_summary.txt`. MAGeCK RRA will output both the negative score and
positive score for each gene. A smaller score indicates higher gene importance. MAGeCK RRA  will
also output the statistical value for the scores of each gene. Genes that are significantly positively
and negatively selected can be identified based on the p-value or FDR.

The downstream analysis of this function includes identifying positive and negative selection genes,
and performing biological functional category analysis and pathway enrichment analysis of these genes.
}
\examples{
data("rra.gene_summary")
data("rra.sgrna_summary")
\dontrun{
    # Run the FluteRRA pipeline
    FluteRRA(rra.gene_summary, rra.sgrna_summary, prefix="RRA", organism="hsa")
}


}
\seealso{
\code{\link{FluteMLE}}
}
\author{
Wubing Zhang
}
