% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrg.R
\docType{methods}
\name{getOrg}
\alias{getOrg}
\title{Determine the gene annotation package.}
\usage{
getOrg(organism)
}
\arguments{
\item{organism}{Character, KEGG species code, or the common species name, used to determine
the gene annotation package. For all potential values check: data(bods); bods. Default org="hsa",
and can also be "human" (case insensitive).}
}
\value{
A list containing three elements:
\item{organism}{species}
\code{pkg}{annotation package name}
\code{Symbol_Entrez}{a data frame, mapping between gene symbol and entrez id}
}
\description{
Determine the gene annotation package. for specific organism
}
\note{
The source can be found by typing \code{MAGeCKFlute:::getOrg}
or \code{getMethod("getOrg")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/getOrg.R}
Users should find it easy to customize this function.
}
\examples{
ann_pkg = getOrg("human")$pkg
print(ann_pkg)

}
\author{
Wubing Zhang
}
