% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{getOrg}
\alias{getOrg}
\title{Get the kegg code of specific mammalia organism.}
\usage{
getOrg(organism)
}
\arguments{
\item{organism}{Character, KEGG species code, or the common species name.
For all potential values check: data(bods); bods. Default org="hsa",
and can also be "human" (case insensitive).}

\item{update}{Boolean, indicating whether download recent annotation from NCBI.}
}
\value{
A list containing three elements:
\item{org}{species}
\code{pkg}{annotation package name}
}
\description{
Get the kegg code of specific mammalia organism.
}
\examples{
ann = getOrg("human")
print(ann$pkg)

}
\author{
Wubing Zhang
}
