% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{gsGetter}
\alias{gsGetter}
\title{Extract pathway annotation from GMT file.}
\usage{
gsGetter(gmtpath = NULL, type = "All", limit = c(0, Inf),
  organism = "hsa")
}
\arguments{
\item{gmtpath}{The path to customized gmt file.}

\item{type}{Geneset category for testing.}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets to load.}

\item{organism}{'hsa' or 'mmu'.}
}
\value{
A three-column data frame.
}
\description{
Extract pathway annotation from GMT file.
}
\examples{
gene2path = gsGetter()
head(gene2path)

}
\author{
Wubing Zhang
}
