% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{gsGetter}
\alias{gsGetter}
\title{Extract pathway annotation from GMT file.}
\usage{
gsGetter(gmtpath = NULL, type = "All", limit = c(0, Inf),
  organism = "hsa", update = FALSE)
}
\arguments{
\item{gmtpath}{The path to customized gmt file.}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (KEGG, REACTOME, C2_CP:PID, C2_CP:BIOCARTA), GO (GOBP, GOCC, GOMF),
MSIGDB (C1, C2 (C2_CP (C2_CP:PID, C2_CP:BIOCARTA), C2_CGP),
C3 (C3_MIR, C3_TFT), C4 (C4_CGN, C4_CM), C5 (C5_BP, C5_CC, C5_MF), C6, C7, H)
and Complex (CORUM). Any combination of them are also accessible
(e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets to load.}

\item{organism}{'hsa' or 'mmu'.}

\item{update}{Boolean, indicating whether update the gene sets from source database.}
}
\value{
A three-column data frame.
}
\description{
Extract pathway annotation from GMT file.
}
\examples{
gene2path = gsGetter(type = "REACTOME+CORUM")
head(gene2path)

}
\author{
Wubing Zhang
}
