% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustView.R
\docType{methods}
\name{hclustView}
\alias{hclustView}
\title{Cluster and view cluster tree}
\usage{
hclustView(d, method = "average", label_cols = NULL, bar_cols = NULL,
  main = NA, xlab = NA, horiz = TRUE, ...)
}
\arguments{
\item{d}{A dissimilarity structure as produced by dist.}

\item{method}{The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{label_cols}{A vector to be used as label's colors for the dendrogram.}

\item{bar_cols}{Either a vector or a matrix, which will be plotted as a colored bar.}

\item{main}{As in 'plot'.}

\item{xlab}{As in 'plot'.}

\item{horiz}{Logical indicating if the dendrogram should be drawn horizontally or not.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Just plot figure in device.
}
\description{
Cluster and view cluster tree
}
\note{
See the vignette for an example of hclustView
Note that the source code of \code{hclustView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::hclustView}
or \code{getMethod("hclustView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/hclustView.R}
Users should find it easy to customize this function.
}
\examples{
label_cols = rownames(USArrests)
hclustView(dist(USArrests), label_cols=label_cols, bar_cols=label_cols)

}
\author{
Wubing Zhang
}
