% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentBarVIew.R
\docType{methods}
\name{IdentBarView}
\alias{IdentBarView}
\title{Identical bar plot}
\usage{
IdentBarView(gg, x = "x", y = "y", fill = c("#CF3C2B", "#394E80"),
  main = NULL, xlab = NULL, ylab = NULL, filename = NULL, width = 5,
  height = 4, ...)
}
\arguments{
\item{gg}{A data frame.}

\item{x}{A character, indicating column (in countSummary) of x-axis.}

\item{y}{A character, indicating column (in countSummary) of y-axis.}

\item{fill}{A character, indicating fill color of all bars.}

\item{main}{A charater, specifying the figure title.}

\item{xlab}{A character, specifying the title of x-axis.}

\item{ylab, }{A character, specifying the title of y-axis.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Identical bar plot
}
\examples{
gg = data.frame(Label = c("Day0_R1", "Day0_R2", "Day7_R1", "Day7_R2"),
                          Reads = c(62818064, 47289074, 51190401, 58686580))
gg$Reads = gg$Reads / sum(gg$Reads)
IdentBarView(gg, x="Label", y="Reads")

}
\author{
Wubing Zhang
}
