\name{loginfo}
\alias{loginfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Log message
}
\description{
Print a log message in console
}
\usage{
loginfo(msg)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{msg}{
A character
}
}

\author{
Wubing Zhang, Feizhen Wu, Binbin Wang
}

\examples{
loginfo("Test the loginfo function ...")
}

