% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAView.R
\docType{methods}
\name{MAView}
\alias{MAView}
\title{MAplot of gene beta scores}
\usage{
MAView(
  beta,
  ctrlname = "Control",
  treatname = "Treatment",
  main = NULL,
  show.statistics = TRUE,
  add.smooth = TRUE,
  lty = 1,
  smooth.col = "red",
  plot.method = c("loess", "lm", "glm", "gam"),
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{beta}{Data frame, including \code{ctrlname} and \code{treatname} as columns.}

\item{ctrlname}{Character vector, specifying the name of control sample.}

\item{treatname}{Character vector, specifying the name of treatment sample.}

\item{main}{As in plot.}

\item{show.statistics}{Show statistics .}

\item{add.smooth}{Whether add a smooth line to the plot.}

\item{lty}{Line type for smooth line.}

\item{smooth.col}{Color of smooth line.}

\item{plot.method}{A string specifying the method to fit smooth line, which should be one of "loess" (default), "lm", "glm" and "gam".}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
MAplot of gene beta scores in Control vs Treatment
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
MAView(dd, ctrlname = "Pmel1_Ctrl", treatname = "Pmel1")
dd2 = NormalizeBeta(dd, method="loess", org = "mmu")
MAView(dd2, ctrlname = "Pmel1_Ctrl", treatname = "Pmel1")

}
\author{
Wubing Zhang
}
