% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAView.R
\docType{methods}
\name{MAView}
\alias{MAView}
\title{MAplot of gene beta scores}
\usage{
MAView(beta, ctrlname = "Control", treatname = "Treatment",
  main = "Negative control normalized", subset = sample(1:length(M),
  min(c(10000, length(M)))), show.statistics = TRUE, span = 2/3,
  family.loess = "gaussian", cex = 1, cex.lab = 1.2, cex.axis = 1,
  cex.main = 1.2, plot.method = c("normal", "smoothScatter", "add"),
  add.loess = TRUE, lwd = 1, lty = 1, loess.col = "red",
  filename = NULL)
}
\arguments{
\item{beta}{Data frame, including \code{ctrlname} and \code{treatname} as columns.}

\item{ctrlname}{Character vector, specifying the name of control sample.}

\item{treatname}{Character vector, specifying the name of treatment sample.}

\item{main}{As in "ma.plot".}

\item{subset}{As in "ma.plot".}

\item{show.statistics}{As in "ma.plot".}

\item{span}{As in "ma.plot".}

\item{family.loess}{As in "ma.plot".}

\item{cex}{As in "ma.plot".}

\item{cex.lab}{As in "ma.plot".}

\item{cex.axis}{As in "ma.plot".}

\item{cex.main}{As in "ma.plot".}

\item{plot.method}{As in "ma.plot".}

\item{add.loess}{As in "ma.plot".}

\item{lwd}{As in "ma.plot".}

\item{lty}{As in "ma.plot".}

\item{loess.col}{As in "ma.plot".}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
no output.}
}
\value{
plot on current device
}
\description{
MAplot of gene beta scores in Control vs Treatment
}
\note{
See the vignette for an example of MAView.
Note that the source code of \code{MAView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::MAView}
or \code{getMethod("MAView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/MAView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
MAView(dd, ctrlname = "D7_R1", treatname = "PLX7_R1", cex=1)

}
\author{
Wubing Zhang
}
