% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAView.R
\docType{methods}
\name{MAView}
\alias{MAView}
\title{MAplot of gene beta scores}
\usage{
MAView(beta, ctrlname = "Control", treatname = "Treatment",
  main = "Negative control normalized", subset = sample(1:length(M),
  min(c(10000, length(M)))), show.statistics = TRUE, span = 2/3,
  family.loess = "gaussian", cex = 1, cex.lab = 1.2, cex.axis = 1,
  cex.main = 1.2, plot.method = c("normal", "smoothScatter", "add"),
  add.loess = TRUE, lwd = 1, lty = 1, loess.col = "red",
  filename = NULL)
}
\arguments{
\item{beta}{data frame, which has columns of 'Gene', \code{ctrlname} and \code{treatname}.}

\item{ctrlname}{character vector, specifying the name of control sample.}

\item{treatname}{character vector, specifying the name of treatment sample.}

\item{main}{as in "ma.plot".}

\item{subset}{as in "ma.plot".}

\item{show.statistics}{as in "ma.plot".}

\item{span}{as in "ma.plot".}

\item{family.loess}{as in "ma.plot".}

\item{cex}{as in "ma.plot".}

\item{cex.lab}{as in "ma.plot".}

\item{cex.axis}{as in "ma.plot".}

\item{cex.main}{as in "ma.plot".}

\item{plot.method}{as in "ma.plot".}

\item{add.loess}{as in "ma.plot".}

\item{lwd}{as in "ma.plot".}

\item{lty}{as in "ma.plot".}

\item{loess.col}{as in "ma.plot".}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
no output.}
}
\value{
plot on current device
}
\description{
MAplot of gene beta scores in Control vs Treatment
}
\note{
See the vignette for an example of MAView.
Note that the source code of \code{MAView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::MAView}
or \code{getMethod("MAView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/MAView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
MAView(dd, cex=1)

}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{DensityView}}

\code{\link{ViolinView}}   \code{\link{SquareView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{KeggPathwayView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
