% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBeta.R
\docType{methods}
\name{normalize.loess}
\alias{normalize.loess}
\alias{loess.normalize}
\title{normalize.loess}
\usage{
normalize.loess(
  mat,
  subset = sample(1:(dim(mat)[1]), min(c(5000, nrow(mat)))),
  epsilon = 10^-2,
  maxit = 1,
  log.it = FALSE,
  verbose = TRUE,
  span = 2/3,
  family.loess = "symmetric",
  ...
)
}
\arguments{
\item{mat}{A matrix with columns containing the values of the chips to normalize.}

\item{subset}{A subset of the data to fit a loess to.}

\item{epsilon}{A tolerance value (supposed to be a small value - used as a stopping criterion).}

\item{maxit}{Maximum number of iterations.}

\item{log.it}{Logical. If \code{TRUE} it takes the log2 of \code{mat}.}

\item{verbose}{Logical. If \code{TRUE} displays current pair of chip being worked on.}

\item{span}{Parameter to be passed the function \code{\link[stats]{loess}}}

\item{family.loess}{Parameter to be passed the function \code{\link[stats]{loess}}.
\code{"gaussian"} or \code{"symmetric"} are acceptable values for this parameter.}

\item{...}{Any of the options of normalize.loess you would like to modify (described above).}
}
\value{
A matrix similar as \code{mat}.
}
\description{
Loess normalization method.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
beta_loess = normalize.loess(dd[,c("dmso", "plx")])

}
\seealso{
\code{\link{loess}}

\code{\link{NormalizeBeta}}
}
\author{
Wubing Zhang
}
