% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBeta.R
\docType{methods}
\name{NormalizeBeta}
\alias{NormalizeBeta}
\alias{normalizebeta}
\title{Normalize gene beta scores}
\usage{
NormalizeBeta(beta, samples = NULL, method = "cell_cycle", id = 1,
  posControl = NULL)
}
\arguments{
\item{beta}{Data frame.}

\item{samples}{Character vector, specifying the sample names in \emph{beta} columns.
If NULL (default), take all \emph{beta} columns as samples.}

\item{method}{Character, one of 'cell_cycle'(default) and 'loess'.}

\item{id}{A single number giving the column of the table which contains the gene names,
or character string giving the name of the table column containing the gene names.}

\item{posControl}{A character vector, specifying a list of positive control genes.}
}
\value{
A data frame with same format as input data \emph{beta}.
}
\description{
Two normalization methods are available. \code{cell_cycle} method normalizes gene beta scores
 based on positive control genes in CRISPR screening. \code{loess} method normalizes gene
 beta scores using loess.
}
\details{
In CRISPR screens, cells treated with different conditions (e.g., with or without drug)
may have different proliferation rates. So it's necessary to normalize the proliferation rate
based on defined positive control genes among samples. After normalization, the beta scores are
comparable across samples. \code{loess} is another optional normalization method, which is used
to normalize array data before.
}
\examples{
data(mle.gene_summary)
data(Zuber_Essential)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(mle.gene_summary, organism="hsa")
#Cell Cycle normalization
dd_essential = NormalizeBeta(dd, samples=c("dmso", "plx"),
    method="cell_cycle", posControl = Zuber_Essential$GeneSymbol)
head(dd_essential)

#Optional loess normalization
dd_loess = NormalizeBeta(dd, samples=c("dmso", "plx"), method="loess")
head(dd_loess)


}
\author{
Wubing Zhang
}
