% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBeta.R
\docType{methods}
\name{NormalizeBeta}
\alias{NormalizeBeta}
\alias{normalizebeta}
\title{Normalize gene beta scores}
\usage{
NormalizeBeta(beta, samples = NULL, method = "cell_cycle",
  posControl = NULL, minus = 0.2)
}
\arguments{
\item{beta}{Data frame, in which rows are EntrezID, columns are samples.}

\item{samples}{Character vector, specifying the samples in \code{beta} to be normalized.
If NULL (default), normalize beta score of all samples in \code{beta}.}

\item{method}{Character, one of 'cell_cycle'(default) and 'loess'.}

\item{posControl}{A file path or a character vector, specifying a list of gene entrezids as positive
controls used for cell cycle normalization}

\item{minus}{Numeric, scale for cell cycle normalization. Between 0 and 1.}
}
\value{
A data frame with same format as input data \code{beta}.
}
\description{
Two normalization methods are available. \code{cell_cycle} method normalizes gene beta scores
 based on positive control genes in CRISPR screening. \code{loess} method normalizes gene
 beta scores using loess.
}
\details{
In CRISPR screens, cells treated with different conditions (e.g., with or without
drug) may have different proliferation rates. So we defined a list of core essential genes,
which is equally negatively selected between samples with different proliferation rate.
Normalization of gene beta scores is performed using these essential genes. \code{cell_cycle}
in MAGeCKFlute normalizes the beta scores of all genes based on the median beta score of essential genes.
After normalization, the beta scores are comparable across samples. \code{loess} is another
optional normalization method, which is used to normalize array data before.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
tmp = TransGeneID(rownames(dd), "Symbol", "Entrez")
dd = dd[!(duplicated(tmp)|is.na(tmp)), ]
rownames(dd) = tmp[!(duplicated(tmp)|is.na(tmp))]
samples=c("D7_R1", "D7_R2", "PLX7_R1", "PLX7_R2")
#Cell Cycle normalization
dd_essential = NormalizeBeta(dd, samples=samples, method="cell_cycle")
head(dd_essential)

#Optional loess normalization
dd_loess = NormalizeBeta(dd, samples=samples, method="loess")
head(dd_loess)


}
\author{
Wubing Zhang
}
