% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmitCommonEssential.R
\docType{methods}
\name{OmitCommonEssential}
\alias{OmitCommonEssential}
\title{Omit common essential genes based on depmap data}
\usage{
OmitCommonEssential(dd, symbol = "id", lineages = "All", dependency = -0.5)
}
\arguments{
\item{dd}{A data frame.}

\item{symbol}{A character, specifying the column name of gene symbols in the data frame.}

\item{lineages}{A character vector, specifying the lineages used for common essential gene selection.}

\item{dependency}{A numeric, specifying the threshold for common essential gene selection.}
}
\value{
A data frame.
}
\description{
Omit common essential genes based on depmap data
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
dim(gdata)
\dontrun{
  rra.omit = OmitCommonEssential(gdata)
  dim(rra.omit)
}

}
\author{
Wubing Zhang
}
