% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankView.R
\docType{methods}
\name{RankView}
\alias{RankView}
\alias{rankview}
\title{View the rank of gene points}
\usage{
RankView(beta, genelist = c(), top = 20, bottom = 20,
  cutoff = c(-sd(beta$diff), sd(beta$diff)), main = NULL, filename = NULL,
  width = 5, height = 4, ...)
}
\arguments{
\item{beta}{Data frame containing two columns. The first column is label of points, and the second column
includes numeric values to be ranked.}

\item{genelist}{Character vector, specifying genes to be labeled in figure.}

\item{top}{Integer, specifying number of top genes to be labeled.}

\item{bottom}{Integer, specifying number of bottom genes to be labeled.}

\item{cutoff}{A two-length numeric vector, in which first value is bottom cutoff, and second value is top cutoff.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Rank all genes according to beta score deviation, and label top and bottom meaningful genes.
Some other interested genes can be labeled too.
}
\note{
See the vignette for an example of RankView.
Note that the source code of \code{RankView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::RankView}
or \code{getMethod("RankView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/RankView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
dd$diff = dd$PLX7_R1 - dd$D7_R1
RankView(dd[,c("Gene", "diff")])

}
\author{
Wubing Zhang
}
