% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankView.R
\docType{methods}
\name{RankView}
\alias{RankView}
\alias{rankview}
\title{Plot rank of gene points}
\usage{
RankView(beta, genelist = c(), top = 10, bottom = 10,
  cutoff = c(sd(beta$diff), -sd(beta$diff)), main = NULL, filename = NULL)
}
\arguments{
\item{beta}{data frame containing columns of "Gene" and "diff".}

\item{genelist}{character vector, specifying labeled genes besides top and bottom labeled genes.}

\item{top}{integer, specifying top number of genes to be labeled.}

\item{bottom}{integer, specifying bottom number of genes to be labeled.}

\item{cutoff}{numeric, cutoff of \code{diff}}

\item{main}{as in 'plot'.}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
no output.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Rank all genes according to beta score deviation, and label top and bottom meaningful genes.
Some other interested genes can be labeled too.
}
\note{
See the vignette for an example of CellCycleView.
Note that the source code of \code{RankView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::RankView}
or \code{getMethod("RankView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/CellCycleView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
dd$diff = dd$Treatment - dd$Control
RankView(dd[,c("Gene", "diff")])

}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{DensityView}}

\code{\link{ViolinView}}   \code{\link{SquareView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{KeggPathwayView}}

\code{\link{MAView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
