% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBeta.R
\docType{methods}
\name{ReadBeta}
\alias{ReadBeta}
\alias{readbeta}
\title{Read gene beta scores}
\usage{
ReadBeta(gene_summary)
}
\arguments{
\item{gene_summary}{A file path or a data frame, which has columns of 'Gene' and beta score of samples.}
}
\value{
A data frame, in which the first column is ENTREZID, and the later columns are beta score for each samples.
}
\description{
Read gene beta scores from file or data frame
}
\examples{
data(mle.gene_summary)
dd = ReadBeta(mle.gene_summary)
head(dd)

}
\author{
Wubing Zhang
}
