% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBeta.R
\docType{methods}
\name{ReadBeta}
\alias{ReadBeta}
\alias{readbeta}
\title{Read gene beta scores from MAGeCK-MLE results}
\usage{
ReadBeta(gene_summary)
}
\arguments{
\item{gene_summary}{A data frame or a file path to gene summary file generated by MAGeCK-MLE.}
}
\value{
A data frame, whose first column is Gene and other columns are comparisons.
}
\description{
Read gene beta scores from MAGeCK-MLE results
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
head(dd)

}
\author{
Wubing Zhang
}
