% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBeta.R
\docType{methods}
\name{ReadBeta}
\alias{ReadBeta}
\alias{readbeta}
\title{Read gene beta scores}
\usage{
ReadBeta(gene_summary, organism = "hsa")
}
\arguments{
\item{gene_summary}{a file path or a data frame, data frame, which has columns of 'Gene',
\code{ctrlname}.beta and \code{treatname}.beta.}

\item{organism}{character, KEGG species code, or the common species name, used to determine
the gene annotation package. For all potential values check: data(bods); bods. Default org="hsa",
and can also be "human" (case insensitive).}
}
\value{
a data frame including four columns, named "Gene", "Control", "Treatment", and "ENTREZID".
}
\description{
Read gene beta scores from file or data frame
}
\note{
The source can be found by typing \code{MAGeCKFlute:::ReadBeta}
or \code{getMethod("ReadBeta")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/ReadBeta.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
dd = ReadBeta(MLE_Data, organism="hsa")
head(dd)

}
\author{
Wubing Zhang
}
