\name{ReadBeta}
\alias{ReadBeta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read gene summary file
}
\description{
Read gene summary file, which is generated by running MAGeCK MLE. The columns of Gene, ctrlname.beta and treatmentname.beta are available.
}
\usage{
ReadBeta(gene_summary, ctrlName = "Control", treatName = "Treatment", organism = "hsa")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gene_summary}{
A file directory or data frame of gene summary table generated by running MAGeCK MLE
}
  \item{ctrlName}{
Name of Control sample
}
  \item{treatName}{
Name of Treatment sample
}
  \item{organism}{
organism type, can be "hsa"(default) or "mmu"
}
}

\value{
A data frame includes four columns, named Gene, Control, Treatment, and ENTREZID
}

\author{
Wubing Zhang, Feizhen Wu, Binbin Wang
}

\examples{
##Need a gene summary file
gene_summary = BRAF_mle.gene_summary
ctrlName = c("D7_R1", "D7_R2")
treatName = c("PLX7_R1", "PLX7_R2")
#Read beta score from gene summary matrix
dd = ReadBeta(gene_summary, ctrlName = ctrlName, treatName = treatName, organism="hsa")
head(dd)
}

