% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\docType{methods}
\name{ReadGMT}
\alias{ReadGMT}
\title{ReadGMT}
\usage{
ReadGMT(gmtpath, limit = c(3, 30))
}
\arguments{
\item{gmtpath}{The path to gmt file.}

\item{limit}{A integer vector of length two, specifying the limit of geneset size.}
}
\value{
An data.frame, in which the first column is gene, and the second column is pathway name.
}
\description{
Parse gmt file to a data.frame
}
\author{
Wubing Zhang
}
