% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRRA.R
\docType{methods}
\name{ReadRRA}
\alias{ReadRRA}
\alias{readrra}
\title{Read gene summary file in MAGeCK-RRA results}
\usage{
ReadRRA(gene_summary, score = c("lfc", "rra")[1])
}
\arguments{
\item{gene_summary}{A file path or a data frame of gene summary data
generated by command `mageck test`.}

\item{score}{lfc or rra, specifying the score type.}
}
\value{
A data frame including three columns, including "id", "LFC" and "FDR".
}
\description{
Read gene summary file in MAGeCK-RRA results
}
\examples{
data(rra.gene_summary)
dd.rra = ReadRRA(rra.gene_summary)
head(dd.rra)

}
\author{
Wubing Zhang
}
