% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRRA.R
\docType{methods}
\name{ReadRRA}
\alias{ReadRRA}
\alias{readrra}
\title{Read MAGeCK-RRA data}
\usage{
ReadRRA(gene_summary, organism = "hsa")
}
\arguments{
\item{gene_summary}{A file path or a data frame, which has three columns named 'id', 'neg.fdr' and 'pos.fdr'.}

\item{organism}{Character, KEGG species code, or the common species name, used to determine
the gene annotation package. For all potential values check: data(bods); bods. Default org="hsa",
and can also be "human" (case insensitive).}
}
\value{
A data frame including four columns, named "Official", "neg.fdr", "pos.fdr" and "ENTREZID".
}
\description{
Read pvalue of gene selection from file or data frame
}
\note{
The source can be found by typing \code{MAGeCKFlute:::ReadRRA}
or \code{getMethod("ReadRRA")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/ReadRRA.R}
Users should find it easy to customize this function.
}
\examples{
data(RRA_Data)
dd.rra = ReadRRA(RRA_Data, organism="hsa")
head(dd.rra)

}
\author{
Wubing Zhang
}
