% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRRA.R
\docType{methods}
\name{ReadRRA}
\alias{ReadRRA}
\alias{readrra}
\title{Read gene summary file in MAGeCK-RRA results}
\usage{
ReadRRA(gene_summary, score = c("lfc", "rra")[1])
}
\arguments{
\item{gene_summary}{A data frame or a file path to gene summary file generated by MAGeCK-RRA.}

\item{score}{"lfc" (default) or "rra", specifying the score type.}
}
\value{
A data frame including three columns, including "id", "LFC" and "FDR".
}
\description{
Read gene summary file in MAGeCK-RRA results
}
\details{
If the score type is equal to lfc, then LFC will be returned. If the score type is rra,
the log10 transformed RRA score will be returned. For FACS-based CRISPR screens, rra score is not recommended.
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
head(gdata)

}
\author{
Wubing Zhang
}
