% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRRA.R
\docType{methods}
\name{ReadRRA}
\alias{ReadRRA}
\alias{readrra}
\title{Read gene summary file in MAGeCK-RRA results}
\usage{
ReadRRA(gene_summary, organism = "hsa")
}
\arguments{
\item{gene_summary}{A file path or a data frame of gene summary data.}

\item{organism}{Character, KEGG species code, or the common species name, used to determine
the gene annotation package. For all potential values check: data(bods); bods. Default org = "hsa",
and can also be "human" (case insensitive).}
}
\value{
A data frame including four columns, named "Official", "EntrezID", "LFC" and "FDR".
}
\description{
Read gene summary file in MAGeCK-RRA results
}
\examples{
data(rra.gene_summary)
dd.rra = ReadRRA(rra.gene_summary, organism = "hsa")
head(dd.rra)

}
\author{
Wubing Zhang
}
