% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadsgRRA.R
\docType{methods}
\name{ReadsgRRA}
\alias{ReadsgRRA}
\title{Read sgRNA summary in MAGeCK-RRA results}
\usage{
ReadsgRRA(sgRNA_summary)
}
\arguments{
\item{sgRNA_summary}{A file path or a data frame of sgRNA summary data.}
}
\value{
A data frame.
}
\description{
Read sgRNA summary in MAGeCK-RRA results
}
\examples{
file2 = file.path(system.file("extdata", package = "MAGeCKFlute"),
                  "testdata/rra.sgrna_summary.txt")
sgrra = ReadsgRRA(file2)
head(sgrra)

}
\author{
Wubing Zhang
}
