% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncorporateDepmap.R
\docType{methods}
\name{ResembleDepmap}
\alias{ResembleDepmap}
\title{Compute the similarity between customized CRISPR screen with Depmap screens}
\usage{
ResembleDepmap(
  dd,
  symbol = "id",
  score = "Score",
  lineages = "All",
  method = c("pearson", "spearman", "kendall")[1]
)
}
\arguments{
\item{dd}{A data frame.}

\item{symbol}{A character, specifying the column name of gene symbols in the data frame.}

\item{score}{A character, specifying the column name of gene essentiality score in the data frame.}

\item{lineages}{A character vector, specifying the lineages used for common essential gene selection.}

\item{method}{A character, indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman".}
}
\value{
A data frame with correlation and test p.value.
}
\description{
Compute the similarity between customized CRISPR screen with Depmap screens
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
\dontrun{
  rra.omit = OmitCommonEssential(gdata)
  depmap_similarity = ResembleDepmap(rra.omit)
  head(depmap_similarity)
}
}
\author{
Wubing Zhang
}
