% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{retrieve_gs}
\alias{retrieve_gs}
\title{Update genesets from source database}
\usage{
retrieve_gs(organism = "hsa")
}
\arguments{
\item{organism}{'hsa' or 'mmu'.}
}
\value{
save data to local library.
}
\description{
Update genesets from source database
}
\examples{
retrieve_gs()

}
\author{
Wubing Zhang
}
