% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterView.R
\docType{methods}
\name{ScatterView}
\alias{ScatterView}
\alias{scatterview}
\title{Scatter plot}
\usage{
ScatterView(beta, ctrlname = "Control", treatname = "Treatment",
  scale_cutoff = 1, main = NULL, filename = NULL, width = 5,
  height = 4, ...)
}
\arguments{
\item{beta}{Data frame, including \code{ctrlname} and \code{treatname} as columns.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the names of treatment samples.}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Scatter plot of all genes, in which x-axis is mean beta score in Control samples, y-axis
is mean beta scores in Treatment samples.
}
\note{
See the vignette for an example of ScatterView.
Note that the source code of \code{ScatterView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::ScatterView}
or \code{getMethod("ScatterView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/ScatterView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
ScatterView(dd, ctrlname = "D7_R1", treatname = "PLX7_R1")


}
\seealso{
\code{\link{SquareView}}
}
\author{
Wubing Zhang
}
