% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterView.R
\docType{methods}
\name{ScatterView}
\alias{ScatterView}
\alias{scatterview}
\title{Scatter plot}
\usage{
ScatterView(beta, ctrlname = "Control", treatname = "Treatment",
  scale_cutoff = 2, main = NULL, filename = NULL, width = 5,
  height = 4, ...)
}
\arguments{
\item{beta}{Data frame, including \code{ctrlname} and \code{treatname} as columns.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the names of treatment samples.}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Scatter plot of all genes, in which x-axis is mean beta score in Control samples, y-axis
is mean beta scores in Treatment samples.
}
\examples{
data(mle.gene_summary)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(mle.gene_summary, organism="hsa")
ScatterView(dd, ctrlname = "dmso", treatname = "plx")


}
\seealso{
\code{\link{SquareView}}
}
\author{
Wubing Zhang
}
