% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRankView.R
\docType{methods}
\name{sgRankView}
\alias{sgRankView}
\title{View sgRNA rank.}
\usage{
sgRankView(
  df,
  gene = NULL,
  top = 3,
  bottom = 3,
  neg_ctrl = NULL,
  binwidth = 0.3,
  interval = 0.1,
  bg.col = "gray90",
  filename = NULL,
  width = 5,
  height = 3.5,
  ...
)
}
\arguments{
\item{df}{A data frame, which contains columns of 'sgrna', 'Gene', and 'LFC'.}

\item{gene}{Character vector, specifying genes to be plotted.}

\item{top}{Integer, specifying number of top genes to be plotted.}

\item{bottom}{Integer, specifying number of bottom genes to be plotted.}

\item{neg_ctrl}{A vector specifying negative ctrl genes.}

\item{binwidth}{A numeric value specifying the bar width.}

\item{interval}{A numeric value specifying the interval length between each bar.}

\item{bg.col}{A character value specifying the background color.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means no output.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}.
}
\description{
View sgRNA rank.
}
\examples{
file2 = file.path(system.file("extdata", package = "MAGeCKFlute"),
                  "testdata/rra.sgrna_summary.txt")
sgrra = ReadsgRRA(file2)
sgRankView(sgrra)

}
\author{
Yihan Xiao
}
