% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareView.R
\docType{methods}
\name{SquareView}
\alias{SquareView}
\alias{squareview}
\title{Scatter plot of 9-Square}
\usage{
SquareView(beta, ctrlname = "Control", treatname = "Treatment",
  label = 0, label.top = TRUE, top = 5, genelist = c(),
  scale_cutoff = 1, main = NULL, filename = NULL, width = 5,
  height = 4, ...)
}
\arguments{
\item{beta}{Data frame, including columns of \code{ctrlname} and \code{treatname}, with Gene Symbol as rowname.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the name of treatment samples.}

\item{label}{An integer or a character specifying the column used as the label, default value is 0 (row names).}

\item{label.top}{Boolean, whether label the top selected genes, default label the top 10 genes in each group.}

\item{top}{Integer, specifying the number of top selected genes to be labeled. Default is 5.}

\item{genelist}{Character vector, specifying labeled genes.}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot a scatter plot with Control beta score as x-axis and Treatment beta score as y-axis,
and colored treatment related genes.
}
\examples{
data(mle.gene_summary)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(mle.gene_summary, organism="hsa")
SquareView(dd, ctrlname = "dmso", treatname = "plx", label = "Gene")


}
\seealso{
\code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
