% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareView.R
\docType{methods}
\name{SquareView}
\alias{SquareView}
\alias{squareview}
\title{Scatter plot of 9-Square}
\usage{
SquareView(
  beta,
  ctrlname = "Control",
  treatname = "Treatment",
  label = 0,
  label.top = TRUE,
  top = 5,
  genelist = c(),
  x_cutoff = NULL,
  y_cutoff = NULL,
  intercept = NULL,
  groups = c("midleft", "topcenter", "midright", "bottomcenter"),
  groupnames = paste0("Group", 1:length(groups)),
  main = NULL,
  filename = NULL,
  width = 6,
  height = 4,
  ...
)
}
\arguments{
\item{beta}{Data frame, including columns of \emph{ctrlname} and \emph{treatname}, with Gene Symbol as rowname.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the name of treatment samples.}

\item{label}{An integer or a character specifying the column used as the label, default value is 0 (row names).}

\item{label.top}{Boolean, whether label the top selected genes, default label the top 10 genes in each group.}

\item{top}{Integer, specifying the number of top selected genes to be labeled. Default is 5.}

\item{genelist}{Character vector, specifying labeled genes.}

\item{x_cutoff}{An one or two-length numeric vector, specifying the cutoff used for x-axis.}

\item{y_cutoff}{An one or two-length numeric vector, specifying the cutoff used for y-axis.}

\item{intercept}{An one or two-length numeric vector, specifying the intercept of diagonal.}

\item{groups}{A character vector, specifying which group to be colored. Optional groups include "topleft",
"topcenter", "topright", "midleft", "midright", "bottomleft", "bottomcenter", "bottomright".}

\item{groupnames}{A character vector, specifying group names.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot a scatter plot with Control beta score as x-axis and Treatment beta score as y-axis,
and colored treatment related genes.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
SquareView(dd, ctrlname = "dmso", treatname = "plx", label = "Gene")


}
\seealso{
\code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
