% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareView.R
\docType{methods}
\name{SquareView}
\alias{SquareView}
\alias{squareview}
\title{Scatter plot of 9-Square}
\usage{
SquareView(beta, ctrlname = "Control", treatname = "Treatment",
  genelist = c(), scale_cutoff = 1.2, main = NULL, filename = NULL)
}
\arguments{
\item{beta}{Data frame, which has columns of 'Gene', \code{ctrlname} and \code{treatname}.}

\item{ctrlname}{A character, specifying the names of control samples.}

\item{treatname}{A character, specifying the name of treatment samples.}

\item{genelist}{Character vector, specifying labeled genes.}

\item{scale_cutoff}{Boolean or numeric, whether scale cutoff to whole genome level,
or how many standard deviation will be used as cutoff.}

\item{main}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plot a scatter plot with Control beta score as x-axis and Treatment beta score as y-axis,
and colored treatment related genes.
}
\note{
See the vignette for an example of SquareView.
Note that the source code of \code{SquareView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::SquareView}
or \code{getMethod("SquareView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/SquareView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
SquareView(dd, ctrlname = "D7_R1", treatname = "PLX7_R1")


}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{DensityView}}

\code{\link{ViolinView}}   \code{\link{MAView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{KeggPathwayView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
