% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{TransGeneID}
\alias{TransGeneID}
\alias{transGeneID}
\title{Gene ID conversion between ENTREZID and SYMBOL}
\usage{
TransGeneID(genes, fromType = "SYMBOL", toType = "ENTREZID",
  organism = "hsa")
}
\arguments{
\item{genes}{A character vector, input genes to be converted.}

\item{fromType}{The input ID type, one of "SYMBOL" (default), "ENTREZID",
"ENSEMBL", "HGNC", "FULLNAME" and "TYPE".}

\item{toType}{The output ID type, one of "SYMBOL", "ENTREZID" (default),
"ENSEMBL", "HGNC", "FULLNAME" and "TYPE".}

\item{organism}{A character, specifying organism, one of "hsa"(or 'Human'),
"mmu"(or 'Mouse'), "bta", "cfa", "ptr", "rno", and "ssc".}
}
\value{
A character vector, named by unique input gene ids.
}
\description{
Gene ID conversion between ENTREZID and SYMBOL
}
\note{
The source can be found by typing \code{MAGeCKFlute:::TransGeneID}
or \code{getMethod("TransGeneID")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/TransGeneID.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
TransGeneID(MLE_Data$Gene[1:10], fromType="SYMBOL", toType="ENTREZID", organism="hsa")

}
\seealso{
\code{\link[pathview]{eg2id}}
}
\author{
Wubing Zhang
}
