% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{TransGeneID}
\alias{TransGeneID}
\alias{transGeneID}
\title{Gene ID conversion between ENTREZID and SYMBOL}
\usage{
TransGeneID(genes, fromType = "Symbol", toType = "Entrez",
  organism = "hsa", fromOrg = organism, toOrg = organism,
  ensemblHost = "www.ensembl.org", update = FALSE)
}
\arguments{
\item{genes}{A character vector, input genes to be converted.}

\item{fromType}{The input ID type, one of "entrez", "symbol"(default), "hgnc",
"ensembl", "fullname" and "uniprotswissprot";
you can also input other valid attribute names for biomaRt.
Look at the code in examples to check valid attributes.}

\item{toType}{The output ID type, similar to `fromType`.}

\item{organism}{"hsa"(default), "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional.}

\item{fromOrg}{"hsa", "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional
(Only used when transform gene ids between organisms).}

\item{toOrg}{"hsa"(default), "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional
(Only used when transform gene ids between organisms).}

\item{ensemblHost}{String, specifying ensembl host, you can use `listEnsemblArchives()`
to show all available Ensembl archives hosts.}

\item{update}{Boolean, specifying whether update built-in gene annotation (needs network and takes time).}
}
\value{
A character vector, named by unique input gene ids.
}
\description{
Gene ID conversion between ENTREZID and SYMBOL
}
\examples{
ds = c("hsapiens_gene_ensembl", "mmusculus_gene_ensembl")
ensembl <- useMart(biomart = 'ENSEMBL_MART_ENSEMBL', dataset = ds)
listAttributes(ensembl)
data(mle.gene_summary)
TransGeneID(mle.gene_summary$Gene[1:10], organism="hsa")
TransGeneID(mle.gene_summary$Gene[1:10], toType="Symbol", fromOrg = "hsa", toOrg = "mmu")
}
\seealso{
\code{\link[pathview]{eg2id}}
}
\author{
Wubing Zhang
}
