% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{TransGeneID}
\alias{TransGeneID}
\alias{transGeneID}
\title{Gene ID conversion between ENTREZID and SYMBOL}
\usage{
TransGeneID(genes, fromType = "Symbol", toType = "Entrez",
  organism = "hsa", useBiomart = FALSE,
  ensemblHost = "www.ensembl.org")
}
\arguments{
\item{genes}{A character vector, input genes to be converted.}

\item{fromType}{The input ID type, one of "Symbol" (default), "Entrez" and "Ensembl";
you can also input other valid attribute names for biomart.}

\item{toType}{The output ID type, one of "Symbol", "Entrez" (default), "Ensembl";
you can also input other valid attribute names for biomart.}

\item{organism}{One of "hsa"(or 'Human'), "mmu"(or 'Mouse'), "bta", "cfa", "ptr", "rno", and "ssc".}

\item{useBiomart}{Boolean, indicating whether use Biomart to do the transformation.}

\item{ensemblHost}{String, specifying ensembl host, you can use `listEnsemblArchives()`
to show all available Ensembl archives hosts.}
}
\value{
A character vector, named by unique input gene ids.
}
\description{
Gene ID conversion between ENTREZID and SYMBOL
}
\examples{
data(mle.gene_summary)
TransGeneID(mle.gene_summary$Gene[1:10], organism="hsa")
TransGeneID(mle.gene_summary$Gene[1:10], organism="hsa", useBiomart = TRUE)

}
\seealso{
\code{\link[pathview]{eg2id}}
}
\author{
Wubing Zhang
}
