% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{TransGeneID}
\alias{TransGeneID}
\title{Gene ID conversion}
\usage{
TransGeneID(
  genes,
  fromType = "Symbol",
  toType = "Entrez",
  organism = "hsa",
  fromOrg = organism,
  toOrg = organism,
  ensemblHost = "www.ensembl.org",
  unique = TRUE,
  update = FALSE
)
}
\arguments{
\item{genes}{A character vector, input genes to be converted.}

\item{fromType}{The input ID type, one of "entrez", "symbol"(default), "hgnc",
"ensembl", "fullname" and "uniprotswissprot";
you can also input other valid attribute names for biomaRt.
Look at the code in examples to check valid attributes.}

\item{toType}{The output ID type, similar to `fromType`.}

\item{organism}{"hsa"(default), "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional.}

\item{fromOrg}{"hsa", "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional
(Only used when transform gene ids between organisms).}

\item{toOrg}{"hsa"(default), "mmu", "bta", "cfa", "ptr", "rno", and "ssc" are optional
(Only used when transform gene ids between organisms).}

\item{ensemblHost}{Character, specifying ensembl host, you can use `listEnsemblArchives()`
to show all available Ensembl archives hosts.}

\item{unique}{Boolean, specifying whether do one-to-one mapping.}

\item{update}{Boolean, specifying whether update built-in gene annotation (needs network and takes time).}
}
\value{
A character vector, named by unique input gene ids.
}
\description{
Gene ID conversion
}
\examples{
TransGeneID("HLA-A", organism="hsa")
TransGeneID("HLA-A", toType = "uniprot", organism="hsa")
TransGeneID("H2-K1", toType="Symbol", fromOrg = "mmu", toOrg = "hsa")

}
\author{
Wubing Zhang
}
