% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinView.R
\docType{methods}
\name{ViolinView}
\alias{ViolinView}
\alias{violinview}
\title{Violin plot}
\usage{
ViolinView(beta, ctrlname = "Control", treatname = "Treatment",
  main = NULL, ylab = "Beta Score", filename = NULL)
}
\arguments{
\item{beta}{data frame, which has columns of 'Gene', \code{ctrlname} and \code{treatname}.}

\item{ctrlname}{character, specifying the name of control sample.}

\item{treatname}{character, specifying the name of treatment sample.}

\item{main}{as in 'plot'.}

\item{ylab}{as in 'plot'.}

\item{filename}{figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plots the violin of beta scores in Control and Treatment samples.
}
\note{
See the vignette for an example of density plot for beta score deviation.
Note that the source code of \code{ViolinView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::ViolinView}
or \code{getMethod("ViolinView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/ViolinView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, ctrlName = "D7_R1", treatName = "PLX7_R1", organism="hsa")
ViolinView(dd)


}
\seealso{
\code{\link{DensityDiffView}}   \code{\link{DensityView}}

\code{\link{MAView}}   \code{\link{SquareView}}

\code{\link{CellCycleView}}  \code{\link{EnrichedView}}

\code{\link{EnrichedGSEView}}  \code{\link{KeggPathwayView}}

\code{\link{RankView}}    \code{\link{ScatterView}}
}
\author{
Wubing Zhang
}
