% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinView.R
\docType{methods}
\name{ViolinView}
\alias{ViolinView}
\alias{violinview}
\title{Violin plot}
\usage{
ViolinView(
  dat,
  samples = NULL,
  main = NULL,
  ylab = "Score",
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{dat}{A data frame.}

\item{samples}{A character vector, specifying the columns in the \code{dat} for plotting.}

\item{main}{A character, specifying title.}

\item{ylab}{A character, specifying title of y-axis.}

\item{filename}{A character, specifying a file name to create on disk.
Set filename to be "NULL", if don't want to save the figure.}

\item{width}{Numeric, specifying width of figure.}

\item{height}{Numeric, specifying height of figure.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Violin plot showing the distribution of numeric vectors with the same length.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
ViolinView(dd[, -1])


}
\seealso{
\code{\link{DensityView}}
}
\author{
Wubing Zhang
}
