% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinView.R
\docType{methods}
\name{ViolinView}
\alias{ViolinView}
\alias{violinview}
\title{Violin plot}
\usage{
ViolinView(
  beta,
  samples = NULL,
  main = NULL,
  ylab = "Beta Score",
  filename = NULL,
  width = 5,
  height = 4,
  ...
)
}
\arguments{
\item{beta}{Data frame, , including \code{samples} as columns.}

\item{samples}{Character, specifying the name of samples to be compared.}

\item{main}{As in 'plot'.}

\item{ylab}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plots the violin of beta scores in Control and Treatment samples.
}
\examples{
file3 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/mle.gene_summary.txt")
dd = ReadBeta(file3)
ViolinView(dd, samples=c("dmso", "plx"))
#or
ViolinView(dd[, c("dmso", "plx")])


}
\seealso{
\code{\link{DensityView}}
}
\author{
Wubing Zhang
}
