% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinView.R
\docType{methods}
\name{ViolinView}
\alias{ViolinView}
\alias{violinview}
\title{Violin plot}
\usage{
ViolinView(beta, samples = NULL, main = NULL, ylab = "Beta Score",
  filename = NULL, width = 5, height = 4, ...)
}
\arguments{
\item{beta}{Data frame, , including all \code{samples} as columns.}

\item{samples}{Character, specifying the name of samples to be compared.}

\item{main}{As in 'plot'.}

\item{ylab}{As in 'plot'.}

\item{filename}{Figure file name to create on disk. Default filename="NULL", which means
don't save the figure on disk.}

\item{width}{As in ggsave.}

\item{height}{As in ggsave.}

\item{...}{Other available parameters in function 'ggsave'.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Plots the violin of beta scores in Control and Treatment samples.
}
\note{
See the vignette for an example of density plot for beta score deviation.
Note that the source code of \code{ViolinView} is very simple.
The source can be found by typing \code{MAGeCKFlute:::ViolinView}
or \code{getMethod("ViolinView")}, or
browsed on github at \url{https://github.com/WubingZhang/MAGeCKFlute/tree/master/R/ViolinView.R}
Users should find it easy to customize this function.
}
\examples{
data(MLE_Data)
# Read beta score from gene summary table in MAGeCK MLE results
dd = ReadBeta(MLE_Data, organism="hsa")
ViolinView(dd, samples=c("D7_R1", "D7_R2", "PLX7_R1", "PLX7_R2"))
#or
ViolinView(dd[, 3:6])


}
\seealso{
\code{\link{DensityView}}
}
\author{
Wubing Zhang
}
