% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoView.R
\docType{methods}
\name{VolcanoView}
\alias{VolcanoView}
\title{Volcano View}
\usage{
VolcanoView(
  df,
  x = "logFC",
  y = "adj.P.Val",
  Label = NA,
  top = 5,
  topnames = NULL,
  x_cutoff = log2(1.5),
  y_cutoff = 0.05,
  mycolour = c("gray80", "#e41a1c", "#377eb8"),
  alpha = 0.6,
  force = 0.1,
  main = NULL,
  xlab = "Log2 Fold Change",
  ylab = "-Log10(Adjust.P)",
  filename = NULL,
  width = 4,
  height = 2.5,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{x}{Colname of df specifying x-axis in Volcanno figure, 'logFC' (default).}

\item{y}{Colname of df specifying y-axis in Volcanno figure, 'adj.P.Val' (default),
which will be plot after log10 transformation.}

\item{Label}{Colname of df specifying labeled terms in Volcanno figure.}

\item{top}{Interger, the number of top significant terms to be labeled.}

\item{topnames}{Character vector, indicating interested terms to be labeled.}

\item{x_cutoff}{Cutoff of x-axis.}

\item{y_cutoff}{Cutoff of y-axis.}

\item{mycolour}{A color vector, specifying colors of non-significant, significant up and down-regulated genes.}

\item{alpha}{Parameter in ggplot.}

\item{force}{Parameter for geom_text_repel.}

\item{main}{Title of volcano figure.}

\item{xlab}{Label of x-axis in figure.}

\item{ylab}{Label of y-axis in figure.}

\item{filename}{Figure file name to create on disk. Default filename="NULL",
which means don't save the figure on disk.}

\item{width}{Width of figure.}

\item{height}{Height of figure.}

\item{...}{Other available parameters in ggsave.}
}
\value{
An object created by \code{ggplot}, which can be assigned and further customized.
}
\description{
Volcano plot
}
\examples{
file1 = file.path(system.file("extdata", package = "MAGeCKFlute"),
"testdata/rra.gene_summary.txt")
gdata = ReadRRA(file1)
VolcanoView(gdata, x = "Score", y = "FDR", Label = "id")

}
\author{
Wubing Zhang
}
